### 2023-07-28 #598 @bozaah
{weatherOz}: an API Client for Australian Weather and Climate Data Resources
Submitting Author Name: Rodrigo Pires

Author Github Handle: @bozaah

Other Package Authors Github handles:@adamhsparks, @arhepworth, @jimii19, @syntrix, @becoleary, @HughParsonage, @PaulMelloy, @jonocarroll, @deanmarchiori, @jimjam-slam, @mpadge, @Keith-Pembleton, @maelle

Repository: https://github.com/DPIRD-FSI/weatherOz

Version submitted: v0.0.1

Submission type: Standard

Version accepted: TBD

```
Package: weatherOz
Title: An API Client for Australian Weather and Climate Data Resources
Version: 0.0.1
  person(
    given = "Rodrigo",
    family = "Pires",
    email = "rodrigo.pires@dpird.wa.gov.au",
    comment = c(ORCID = "0000-0001-7384-6849")
  ),
  person(
    given = "Anna",
    family = "Hepworth",
    role = "aut",
    email = "anna.hepworth@dpird.wa.gov.au",
    comment = c(ORCID = "0000-0003-0204-6347")
  ),
  person(
    given = "Rebecca",
    family = "O'Leary",
    role = "aut",
    email = "bec.oleary@curtin.edu.au"
  ),
  person(
    given = "Jonathan",
    family = "Carroll",
    role = "aut",
    email = "rpkg@jcarroll.com.au",
    comment = c(ORCID = "0000-0002-1404-5264")
  ),
  person(
    given = "James",
    family = "Goldie",
    role = "aut",
    email = "me@jamesgoldie.dev",
    comment = c(ORCID = "0000-0002-5024-6207")
  ),
  person(
    given = "Dean",
    family = "Marchiori",
    role = "aut",
    email = "deanmarchiori@gmail.com"
  ),
  person(
    given = "Paul",
    family = "Melloy",
    role = "aut",
    email = "paul@melloy.com.au",
    comment = c(ORCID = "0000-0003-4253-7167")
  ),
  person(
    given = "Mark",
    family = "Padgham",
    role = "aut",
    email = "mark.padgham@email.com",
    comment = c(ORCID = "0000-0003-2172-5265")
  ),
  person(
    given = "Hugh",
    family = "Parsonage",
    role = "aut",
    email = "hugh.parsonage@gmail.com",
    comment = c(ORCID = "0000-0003-4055-0835")
  ),
  person(
    given = "Keith",
    family = "Pembleton",
    role = "aut",
    email = "keith.pembleton@usq.edu.au",
    comment = c(ORCID = "0000-0002-1896-4516")
  ),
  person(
    given = "Maëlle",
    family = "Salmon",
    role = "ctb",
    email = "maelle.salmon@yahoo.se",
    comment = c(ORCID = "0000-0002-2815-0399",
                "Contributed to debugging a nasty little bug with CI where
                timezones caused tests to fail due to vcr not recognising the
                URL when run outside of Australia/Perth TZ! Suggested the use of
                `withr::local_timzeone()`, see
                <https://github.com/DPIRD-FSI/weatherOz/commit/b052bf91973b8d7e147a39e8938405a64622634b>.")),
  person(
    given = "Jimmy",
    family = "Ng",
    role = "ctb",
    email = "jimmy.ng@dpird.wa.gov.au"
  ),
  person(
    given = "Steve",
    family = "Collins",
    role = "ctb",
    email = "steve.collins@dpird.wa.gov.au",
    comment = c("Designed weatherOz's hex logo.")
  ),
  person(
    given = "Adam H.",
    family = "Sparks",
    email = "adam.sparks@dpird.wa.gov.au",
    role = "aut",
    comment = c(ORCID = "0000-0002-0061-8359")
  ),
  person(given = "Western Australia Agriculture Authority (WAAA)",
         role = c("cph")))
Description: Provides automated downloading, parsing and formatting of weather
    data for Australia through API endpoints provided by the Department of
    Primary Industries and Regional Development (DPIRD) of Western Australia
    and by the Science and Technology Division of the Queensland Government's
    Department of Environment and Science (DES).  As well as the Bureau of
    Meteorology (BOM) of the Australian government precis and coastal
    forecasts, agriculture bulletin data, and downloading and importing radar
    and satellite imagery files.  DPIRD weather data are accessed through public
    APIs provided by DPIRD, <https://www.agric.wa.gov.au/weather-api-20>,
    providing access to weather station data from DPIRD's weather station
    network.  Australia-wide weather data are based on data from the Australian
    Bureau of Meteorology (BOM) data and accessed through SILO (Scientific
    Information for Land Owners) (Jeffery et al., 2001,
    <doi: 10.1016/S1364-8152(01)00008-1>).  DPIRD data are made available under
    a Creative Commons Attribution 3.0 Licence (CC BY 3.0 AU) license
    <https://creativecommons.org/licenses/by/3.0/au/deed.en>.  SILO data are
    released under a  Creative Commons Attribution 4.0 International licence
    (CC BY 4.0) <https://creativecommons.org/licenses/by/4.0/>.  BOM data are
    (c) Australian Government Bureau of Meteorology and released under a
    Creative Commons (CC) Attribution 3.0 licence or Public Access Licence (PAL)
    as appropriate, see <http://www.bom.gov.au/other/copyright.shtml> for
    further details.
    apsimx,
    clock,
    crayon,
    crul,
    curl,
    foreign,
    grDevices,
    magick,
    readr (>= 1.1.1),
    stringi,
    stars,
    terra,
    xml2
    covr,
    ggthemes,
    grid,
    gridExtra,
    mapproj,
    maps,
    mailR,
    vcr (>= 0.6.0),
    vdiffr,
    withr
X-schema.org-applicationCategory: Tools
X-schema.org-keywords: dpird, bom, meteorological-data, weather-forecast,
    australia, weather, weather-data, meteorology, western-australia,
    australia-bureau-of-meteorology, western-australia-agriculture
X-schema.org-isPartOf: https://www.agric.wa.gov.au/
```

-   Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories)
    this package falls under: (Please check an appropriate box below. If
    you are unsure, we suggest you make a pre-submission inquiry.):

-   Explain how and why the package falls under these categories
    (briefly, 1-2 sentences)

    {weatherOz} aims to facilitate access and download weather and
    climate data for Australia from Australian data sources. Data are
    sourced from from the Western Australian Department of Primary
    Industries and Regional Development (DPIRD) and the Scientific
    Information for Land Owners (SILO) API endpoints and the Australian
    Government Bureau of Meteorology's (BOM) FTP server.

-   Who is the target audience and what are scientific applications of
    this package?

    The target audience for this package is researchers, industry
    stakeholders and endusers who need to find, access and download and
    analyse weather and climate data from Australian data sources.

-   Are there other R packages that accomplish the same thing? If so,
    how does yours differ or meet [our criteria for
    best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)?

    No available package has the functionality and capabilities to find,
    access and download weather and climate data for Australia from
    multiple Australian data sources.

    The {bomrang} package, currently archived and unsupported due to
    Australian Government Bureau of Meteorology (BOM) policy changes,
    has functionality to source data from BOM. Some of the existing
    functionality that is permitted by the BOM has been incorporated
    into {weatherOz}.

-   (If applicable) Does your package comply with our [guidance around
    *Ethics, Data Privacy and Human Subjects
    Research*](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

-   If you made a pre-submission inquiry, please paste the link to the
    corresponding issue, forum post, or other discussion, or @tag the
    editor you contacted.

    No pre-submission inquiry submitted.

-   Explain reasons for any [`pkgcheck`
    items](https://docs.ropensci.org/pkgcheck/) which your package is
    unable to pass.

    Currently the {weatherOz} passes all `pkgcheck` items.

-   [x] I have read the [rOpenSci packaging
    guide](https://devguide.ropensci.org/building.html).
-   [x] I have read the [author
    guide](https://devdevguide.netlify.app/authors-guide.html) and I
    expect to maintain this package for at least 2 years or to find a
    replacement.

-   [x] does not violate the Terms of Service of any service it
    interacts with.
-   [x] contains a [README with instructions for installing the
    development
    version](https://ropensci.github.io/dev_guide/building.html#readme).
-   [x] includes [documentation with examples for all functions, created
    with
    roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
-   [x] contains a vignette with examples of its essential functions and
    uses.
-   [x] has a [test
    suite](https://ropensci.github.io/dev_guide/building.html#testing).
-   [x] has [continuous
    integration](https://ropensci.github.io/dev_guide/ci.html),
    including reporting of test coverage.

-   [ ] Do you wish to submit an Applications Article about your package
    to [Methods in Ecology and
    Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)?
    If so:

-   [ ] The package is novel and will be of interest to the broad
    readership of the journal.
-   [ ] The manuscript describing the package is no longer than 3000
    words.
-   [ ] You intend to archive the code for the package in a long-term
    repository which meets the requirements of the journal (see [MEE's
    Policy on Publishing
    Code](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
-   (*Scope: Do consider MEE's [Aims and
    Scope](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html)
    for your manuscript. We make no guarantee that your manuscript will
    be within MEE scope.*)
-   (*Although not required, we strongly recommend having a full
    manuscript prepared when you submit here.*)
-   (*Please do not submit your package separately to Methods in Ecology
    and Evolution*)

-   [x] I agree to abide by [rOpenSci's Code of
    Conduct](https://ropensci.org/code-of-conduct/) during the review
    process and in maintaining my package should it be accepted.
