### 2023-07-28 4b94df5
Update NEWS.md with first bugfix!
