+++ NEWS.md
@@ -1,3 +1,9 @@
-# weatherOz 0.0.0.9000
+# weatherOz 0.0.1.9000
+
+## Bug fixes
+
+* Fixed [bug](https://github.com/DPIRD-FSI/weatherOz/issues/35) where `find_forecast_town()` returned all values, not the proper values.
+
+# weatherOz 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
