+++ man/get_dpird_extremes.Rd
@@ -111,7 +111,7 @@ xtreme <- get_dpird_extremes(
   station_code = "BR",
   type = c("erosionCondition",
            "heatCondition"),
-  api_key = my_key
+  api_key = "your_api_key"
 )
 }
 
