+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Fixed [bug](https://github.com/DPIRD-FSI/weatherOz/issues/35) where `find_forecast_town()` returned all values, not the proper values.
 
+* Fixed example in `get_dpird_extremes()` that didn't use "your_api_key" as all other examples did for error checking.
+
 # weatherOz 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
