### 2023-07-28 d03e683
handle terra::plot() re-export more elegantly
