+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -158,41 +158,41 @@ library(weatherOz)
   )
 )
 #>       station_code           date_time air_temperature relative_humidity soil_temperature
-#>    1:           NO 2023-07-26 20:06:00             8.5              99.2             10.7
-#>    2:           NO 2023-07-26 20:07:00             8.7              99.2             10.7
-#>    3:           NO 2023-07-26 20:08:00             8.7              99.2             10.7
-#>    4:           NO 2023-07-26 20:09:00             8.7              99.2             10.6
-#>    5:           NO 2023-07-26 20:10:00             8.6              99.2             10.6
+#>    1:           NO 2023-07-28 08:05:00            11.0              85.7              9.6
+#>    2:           NO 2023-07-28 08:06:00            11.1              85.6              9.6
+#>    3:           NO 2023-07-28 08:07:00            11.0              85.8              9.6
+#>    4:           NO 2023-07-28 08:08:00            11.0              85.8              9.7
+#>    5:           NO 2023-07-28 08:09:00            11.0              86.0              9.7
 #>   ---                                                                                    
-#> 1435:           NO 2023-07-27 20:00:00            11.2              83.0             11.6
-#> 1436:           NO 2023-07-27 20:01:00            11.1              83.4             11.5
-#> 1437:           NO 2023-07-27 20:02:00            11.1              83.6             11.5
-#> 1438:           NO 2023-07-27 20:03:00            11.1              83.9             11.5
-#> 1439:           NO 2023-07-27 20:04:00            11.1              84.4             11.5
+#> 1432:           NO 2023-07-29 07:56:00            11.3              99.2              9.8
+#> 1433:           NO 2023-07-29 07:57:00            11.4              99.2              9.8
+#> 1434:           NO 2023-07-29 07:58:00            11.4              99.2              9.9
+#> 1435:           NO 2023-07-29 07:59:00            11.5              99.2              9.9
+#> 1436:           NO 2023-07-29 08:00:00            11.5              99.2              9.9
 #>       solar_irradiance rainfall dew_point wet_bulb wind_height wind_avg_speed
-#>    1:                1        0       8.4      8.4           3           1.04
-#>    2:                1        0       8.6      8.6           3           0.86
-#>    3:                2        0       8.6      8.6           3           0.40
-#>    4:                2        0       8.6      8.6           3           0.97
-#>    5:                2        0       8.5      8.5           3           0.04
+#>    1:               79        0       8.7      9.9           3           6.16
+#>    2:               84        0       8.8     10.0           3           5.15
+#>    3:               88        0       8.7     10.0           3           7.06
+#>    4:               94        0       8.7     10.0           3           6.23
+#>    5:              100        0       8.8     10.0           3           5.04
 #>   ---                                                                        
-#> 1435:                1        0       8.4      9.9           3           0.04
-#> 1436:                1        0       8.4      9.9           3           0.00
-#> 1437:                1        0       8.4      9.9           3           0.00
-#> 1438:                2        0       8.5      9.9           3           0.00
-#> 1439:                2        0       8.6      9.9           3           0.00
+#> 1432:               89        0      11.2     11.2           3           6.55
+#> 1433:              117        0      11.3     11.3           3           5.54
+#> 1434:              118        0      11.3     11.3           3           4.97
+#> 1435:               96        0      11.4     11.4           3           3.78
+#> 1436:              101        0      11.4     11.4           3           4.39
[…]
