+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -158,41 +158,41 @@ library(weatherOz)
   )
 )
 #>       station_code           date_time air_temperature relative_humidity soil_temperature
-#>    1:           NO 2023-07-28 08:05:00            11.0              85.7              9.6
-#>    2:           NO 2023-07-28 08:06:00            11.1              85.6              9.6
-#>    3:           NO 2023-07-28 08:07:00            11.0              85.8              9.6
-#>    4:           NO 2023-07-28 08:08:00            11.0              85.8              9.7
-#>    5:           NO 2023-07-28 08:09:00            11.0              86.0              9.7
+#>    1:           NO 2023-07-28 08:16:00            11.1              86.3              9.7
+#>    2:           NO 2023-07-28 08:17:00            11.1              86.2              9.7
+#>    3:           NO 2023-07-28 08:18:00            11.1              86.0              9.7
+#>    4:           NO 2023-07-28 08:19:00            11.1              85.8              9.7
+#>    5:           NO 2023-07-28 08:20:00            11.1              85.5              9.7
 #>   ---                                                                                    
-#> 1432:           NO 2023-07-29 07:56:00            11.3              99.2              9.8
-#> 1433:           NO 2023-07-29 07:57:00            11.4              99.2              9.8
-#> 1434:           NO 2023-07-29 07:58:00            11.4              99.2              9.9
-#> 1435:           NO 2023-07-29 07:59:00            11.5              99.2              9.9
-#> 1436:           NO 2023-07-29 08:00:00            11.5              99.2              9.9
+#> 1431:           NO 2023-07-29 08:06:00            11.7              99.2              9.9
+#> 1432:           NO 2023-07-29 08:07:00            11.8              99.2              9.9
+#> 1433:           NO 2023-07-29 08:08:00            11.9              99.2              9.9
+#> 1434:           NO 2023-07-29 08:09:00            12.0              99.2             10.0
+#> 1435:           NO 2023-07-29 08:10:00            12.1              99.2             10.0
 #>       solar_irradiance rainfall dew_point wet_bulb wind_height wind_avg_speed
-#>    1:               79        0       8.7      9.9           3           6.16
-#>    2:               84        0       8.8     10.0           3           5.15
-#>    3:               88        0       8.7     10.0           3           7.06
-#>    4:               94        0       8.7     10.0           3           6.23
-#>    5:              100        0       8.8     10.0           3           5.04
+#>    1:              106        0       8.9     10.1           3           5.65
+#>    2:              109        0       8.9     10.1           3           6.19
+#>    3:              115        0       8.8     10.1           3           7.34
+#>    4:              120        0       8.8     10.1           3           7.81
+#>    5:              120        0       8.8     10.0           3          10.76
 #>   ---                                                                        
-#> 1432:               89        0      11.2     11.2           3           6.55
-#> 1433:              117        0      11.3     11.3           3           5.54
-#> 1434:              118        0      11.3     11.3           3           4.97
-#> 1435:               96        0      11.4     11.4           3           3.78
-#> 1436:              101        0      11.4     11.4           3           4.39
+#> 1431:              104        0      11.6     11.6           3           5.04
+#> 1432:               93        0      11.7     11.7           3           6.30
+#> 1433:               89        0      11.8     11.8           3           6.95
+#> 1434:              101        0      11.9     11.9           3           5.69
+#> 1435:              118        0      12.0     12.0           3           7.09
 #>       wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>    1:                              WNW                        293          4.176           9.22
-#>    2:                               NW                        310          1.656           9.22
-#>    3:                               NW                        317          4.176          11.74
-#>    4:                              NNW                        345          4.176          11.74
[…]
