+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: weatherOz
 Title: An API Client for Australian Weather and Climate Data Resources
-Version: 0.0.1
+Version: 0.0.1.9000
 Authors@R: c(
   person(
     given = "Rodrigo",
