### 2023-07-29 972538c
Update NEWS and codemeta.json
