+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * Fix the re-export of `terra::plot()`.
 
+* Fix example for `get_dpird_extremes()` that didn't follow the package's standard use of "your_api_key" for the `api_key` value in the example.
+
 # weatherOz 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
