### 2023-07-29 4b92b2f
Cyclocomp for find_nearby_stations at 12 now. Shorten line in tests
