### 2023-07-29 175edf6
clean up logic when filtering values
