### 2023-07-29 5a3f723
Reduce cyclomatic complexity
