### 2023-07-29 dc1a17d
Set up fn to validate shared user-input checks
