### 2023-07-29 0e22cda
Specify return in fn
