### 2023-07-29 f9056a4
Add @noRd to .check_silo_values
