### 2023-07-29 a4ef929
Fix unassigned fn output when checking values
