+++ R/get_dpird_extremes.R
@@ -134,7 +134,7 @@ get_dpird_extremes <- function(station_code,
   .check_not_example_api_key(api_key)
 
   if (any(values != "all") &&
-      any(values %notin% dpird_extreme_weather_values)) {
+      any(values %notin% weatherOz::dpird_extreme_weather_values)) {
     stop(call. = FALSE,
          "You have specified invalid weather values.")
   }
@@ -143,13 +143,14 @@ get_dpird_extremes <- function(station_code,
     .values <-  c("stationCode",
                   "longitude",
                   "latitude",
-                  dpird_extreme_weather_values)
+                  weatherOz::dpird_extreme_weather_values)
   } else {
     .values <-
       c("stationCode",
         "longitude",
         "latitude",
-        dpird_extreme_weather_values[dpird_extreme_weather_values %in% values])
+        weatherOz::dpird_extreme_weather_values[
+          weatherOz::dpird_extreme_weather_values %in% values])
   }
 
   query_list <- list(
