### 2023-07-31 #598 @bozaah
Thank you! We have addressed the✖️ issues: dealt with `terra::plot()` export, reduced complexity of `get_dpird_summaries()` and `find_nearby_stations()`, and we expect that users should be aware of `get_station_metadata()` if loading both weatherOz and stationaRy. Users can use `::` to select a function from the intended package and/or utilities such as the {conflicted} package.
