### 2023-08-07 336fabd
Closes https://github.com/DPIRD-FSI/weatherOz/issues/36
