### 2023-08-07 2b026c7
Update NEWS and codemeta.json
