### 2023-08-21 0d4934e
Clarify error message when BOM FTP server is unresponsive
