### 2023-08-23 9a28123
Fix bug when requesting nearest stations and the station_code leading 0s were not padded
