+++ R/query_silo_api.R
@@ -152,8 +152,8 @@
       skip_absent = TRUE
     )
 
-    response_data[, station_code := as.factor(as.character(
-      trimws(sprintf("%06s", station_code))))]
+    response_data[, station_code :=
+                    as.factor(as.character(sprintf("%06d", station_code)))]
     response_data[, station_name := trimws(.strcap(x = station_name))]
     response_data[, owner := "BOM"]
     response_data[, distance_km := round(distance_km, 1)]
