### 2023-08-23 1b05d93
Fix another integer/character sprintf issue with leading 0s
