### 2023-08-23 bb13ef8
Update NEWS.md with bugfix
