### 2023-08-29 e39b16d
Should fix issue with truncated DPIRD summaries
