### 2023-08-29 a6e6324
Reknit vignettes to ensure quality after bug fix
