+++ vignettes/weatherOz_for_BOM.Rmd
@@ -64,7 +64,6 @@ library(weatherOz)
 #>   Please note information at the page 
 #>   <http://www.bom.gov.au/catalogue/data-feeds.shtml#precis>.
 #>   --------------------------------------------------------------------------------  
-<<<<<<< HEAD
 #>      index product_id state         town       aac       lat      lon elev
 #>   1:     0   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
 #>   2:     1   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
@@ -72,107 +71,59 @@ library(weatherOz)
 #>   4:     3   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
 #>   5:     4   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
 #>  ---                                                                      
-#> 787:     2   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 788:     3   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 789:     4   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 790:     5   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 791:     6   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 900:     3   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 901:     4   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 902:     5   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 903:     6   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 904:     7   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
 #>         start_time_local end_time_local utc_offset      start_time_utc
-#>   1: 2023-07-28 09:30:16     2023-07-29      10:00 2023-07-27 23:30:16
-#>   2: 2023-07-29 00:00:00     2023-07-30      10:00 2023-07-28 14:00:00
-#>   3: 2023-07-30 00:00:00     2023-07-31      10:00 2023-07-29 14:00:00
-#>   4: 2023-07-31 00:00:00     2023-08-01      10:00 2023-07-30 14:00:00
-#>   5: 2023-08-01 00:00:00     2023-08-02      10:00 2023-07-31 14:00:00
+#>   1: 2023-08-29 17:00:00     2023-08-30      10:00 2023-08-29 07:00:00
+#>   2: 2023-08-30 00:00:00     2023-08-31      10:00 2023-08-29 14:00:00
+#>   3: 2023-08-31 00:00:00     2023-09-01      10:00 2023-08-30 14:00:00
+#>   4: 2023-09-01 00:00:00     2023-09-02      10:00 2023-08-31 14:00:00
+#>   5: 2023-09-02 00:00:00     2023-09-03      10:00 2023-09-01 14:00:00
 #>  ---                                                                  
-#> 787: 2023-07-30 00:00:00     2023-07-31      10:00 2023-07-29 14:00:00
-#> 788: 2023-07-31 00:00:00     2023-08-01      10:00 2023-07-30 14:00:00
[…]
