+++ vignettes/weatherOz.Rmd
@@ -64,9 +64,12 @@ Northam
 #>    station_code station_name longitude  latitude state elev_m
 #> 1:           NO      Northam  116.6942 -31.65161    WA    163
 #> 2:           MK       Muresk  116.6913 -31.72772    WA    251
-#>                                                                   owner distance_km
-#> 1: WA Department of Primary Industries and Regional Development (DPIRD)        3.23
-#> 2: WA Department of Primary Industries and Regional Development (DPIRD)        7.93
+#>                                                                   owner
+#> 1: WA Department of Primary Industries and Regional Development (DPIRD)
+#> 2: WA Department of Primary Industries and Regional Development (DPIRD)
+#>    distance_km
+#> 1:        3.23
+#> 2:        7.93
 ```
 
 There are two stations within 20km of the coordinates for Northam, WA that we provided.
@@ -95,30 +98,30 @@ dpird_daily
 #> 363:           NO      Northam 2022    12  29 2022-12-29     26.0     40.2
 #> 364:           NO      Northam 2022    12  30 2022-12-30     26.6     39.2
 #> 365:           NO      Northam 2022    12  31 2022-12-31     22.7     31.1
-#>            air_tmax_time air_tmin       air_tmin_time rainfall rh_avg rh_tmax
-#>   1: 2022-01-01 17:02:00     16.7 2022-01-01 05:17:00        0   39.9    71.5
-#>   2: 2022-01-02 16:58:00     14.4 2022-01-02 05:19:00        0   35.2    63.8
-#>   3: 2022-01-03 14:42:00     16.6 2022-01-03 05:18:00        0   38.0    68.5
-#>   4: 2022-01-04 15:15:00     17.3 2022-01-04 04:22:00        0   45.9    74.2
-#>   5: 2022-01-05 16:19:00     16.1 2022-01-05 05:14:00        0   42.1    77.2
-#>  ---                                                                         
-#> 361: 2022-12-27 14:48:20     13.7 2022-12-27 05:13:20        0   41.8    81.1
-#> 362: 2022-12-28 19:58:20     11.9 2022-12-28 05:33:20        0   32.8    65.4
-#> 363: 2022-12-30 03:03:20     13.8 2022-12-29 05:01:40        0   29.2    62.5
-#> 364: 2022-12-30 13:50:50     14.5 2022-12-30 04:52:30        0   34.4    61.2
-#> 365: 2022-12-31 15:45:50     16.2 2022-12-31 05:23:20        0   59.0    98.9
-#>             rh_tmax_time rh_tmin        rh_tmin_time
-#>   1: 2022-01-01 04:10:00    14.5 2022-01-01 18:03:00
-#>   2: 2022-01-02 05:39:00    16.4 2022-01-02 16:58:00
-#>   3: 2022-01-03 05:22:00    14.4 2022-01-03 16:15:00
-#>   4: 2022-01-04 04:33:00    20.1 2022-01-04 16:35:00
-#>   5: 2022-01-05 05:22:00    14.4 2022-01-05 17:21:00
-#>  ---                                                
-#> 361: 2022-12-26 16:03:20    14.6 2022-12-27 16:47:30
-#> 362: 2022-12-28 05:35:00     9.2 2022-12-28 16:11:40
-#> 363: 2022-12-29 05:08:20     9.5 2022-12-29 17:31:40
-#> 364: 2022-12-30 05:06:40     7.9 2022-12-30 14:00:00
-#> 365: 2022-12-31 05:30:00    24.7 2022-12-31 16:01:40
+#>            air_tmax_time air_tmin       air_tmin_time rainfall rh_avg
[…]

[…]
