### 2023-08-29 eba532d
Update NEWS.md with newest bugfix
