### 2023-09-01 cd0debc
Update README to be sure it's using latest version of pkg
