### 2023-09-02 d477b62
Clean up NEWS.md
