### 2023-09-08 a02a287
Use proper api.dpird URL now that IS have fixed the issue with the server response
