### 2023-09-12 #598 @bozaah
Dear @ldecicco-USGS,
Thank you for taking up the editor role for our package. I appreciate yours and @JanLauGe's time working on this.

I am writing to mention that I will be on leave until the first week of November so I will be somehow slow and/or unresponsive to queries from you and the package reviewers. I will address the comments and necessary edits upon my return. Thank you!
