### 2023-09-13 7a9cf92
Fix error message when no SILO stations found nearby lat/lon
