### 2023-09-13 b39ffc2
If there are no combined stations, return invisible NULL
