### 2023-09-13 7657200
Use error when there are no stations found at all
