### 2023-09-16 5a18b2d
Fix bug when no stations are matched nearby in either network
