### 2023-09-16 07d6e65
fix bug when checking for results from both APIs for nearest stations
