### 2023-09-16 3b4c29c
Finally FIXES messages and errors when no nearby stations are found
