### 2023-10-10 #598 @JanLauGe
Apologies for the delay in getting this off the ground. I'd love to kick this off this week if that is still okay with everybody?
