### 2023-10-16 98f93f6
Use space after "<" in message to align DPIRD and SILO msgs
