### 2023-10-23 4919acb
Filter out empty rows from SILO responses that were associated with metadata rows
