### 2023-10-23 7e298ee
Fix bug with incorrect values in month and day fields
