### 2023-10-23 47361fb
Fixes bug in patched point data where station name was NA
