### 2023-11-02 53b8455
Add lon/lat values to DPIRD summaries and update vvignettes
