+++ vignettes/weatherOz_for_SILO.Rmd
@@ -105,114 +105,114 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude year month  day       date air_tmax air_tmax_source
-#>   1:       150      -27 2022  2022 2022 2022-01-01     29.3              25
-#>   2:       150      -27 2022  2022 2022 2022-01-02     32.4              25
-#>   3:       150      -27 2022  2022 2022 2022-01-03     34.6              25
-#>   4:       150      -27 2022  2022 2022 2022-01-04     35.0              25
-#>   5:       150      -27 2022  2022 2022 2022-01-05     35.5              25
-#>  ---                                                                       
-#> 601:       150      -27 2023  2023 2023 2023-08-24     26.3              25
-#> 602:       150      -27 2023  2023 2023 2023-08-25     26.1              25
-#> 603:       150      -27 2023  2023 2023 2023-08-26     26.1              25
-#> 604:       150      -27 2023  2023 2023 2023-08-27     25.8              25
-#> 605:       150      -27 2023  2023 2023 2023-08-28     22.8              75
-#>      air_tmin air_tmin_source  elev_m et_morton_actual
-#>   1:     18.8              25 274.9 m              4.8
-#>   2:     16.8              25 274.9 m              5.0
-#>   3:     18.2              25 274.9 m              5.1
-#>   4:     18.0              25 274.9 m              4.5
-#>   5:     20.4              25 274.9 m              5.3
-#>  ---                                                  
-#> 601:      3.1              25 274.9 m              0.5
-#> 602:      4.3              25 274.9 m              2.7
-#> 603:      6.2              25 274.9 m              2.0
-#> 604:      8.1              25 274.9 m              2.4
-#> 605:      8.6              25 274.9 m              3.2
-#>      et_morton_actual_source et_morton_potential et_morton_potential_source
-#>   1:                      26                 6.0                         26
-#>   2:                      26                 7.6                         26
-#>   3:                      26                10.5                         26
-#>   4:                      26                 9.2                         26
-#>   5:                      26                 9.7                         26
-#>  ---                                                                       
-#> 601:                      26                 5.6                         26
-#> 602:                      26                 3.8                         26
-#> 603:                      26                 3.7                         26
-#> 604:                      26                 4.7                         26
-#> 605:                      26                 4.2                         26
-#>      et_morton_wet et_morton_wet_source et_short_crop et_short_crop_source
-#>   1:           5.4                   26           4.5                   26
-#>   2:           6.3                   26           5.7                   26
-#>   3:           7.8                   26           7.3                   26
-#>   4:           6.9                   26           6.6                   26
-#>   5:           7.5                   26           6.9                   26
-#>  ---                                                                      
-#> 601:           3.0                   26           3.9                   26
-#> 602:           3.2                   26           3.2                   26
-#> 603:           2.9                   26           3.1                   26
-#> 604:           3.5                   26           3.4                   26
-#> 605:           3.7                   26           3.1                   26
-#>      et_tall_crop et_tall_crop_source evap_comb evap_comb_source
-#>   1:          5.3                  26       4.9               25
-#>   2:          7.0                  26       9.6               25
-#>   3:          9.1                  26       8.2               25
-#>   4:          8.3                  26       8.6               25
-#>   5:          8.5                  26       6.3               25
-#>  ---                                                            
-#> 601:          5.7                  26       4.6               25
-#> 602:          4.2                  26       4.5               25
-#> 603:          4.1                  26       5.2               25
[…]
