+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -158,53 +158,53 @@ library(weatherOz)
   )
 )
 #>       station_code           date_time air_temperature relative_humidity
-#>    1:           NO 2023-08-28 19:34:00            16.7              67.5
-#>    2:           NO 2023-08-28 19:35:00            16.7              67.7
-#>    3:           NO 2023-08-28 19:36:00            16.6              67.8
-#>    4:           NO 2023-08-28 19:37:00            16.6              68.0
-#>    5:           NO 2023-08-28 19:38:00            16.6              68.0
+#>    1:           NO 2023-11-01 08:30:00            20.7              47.8
+#>    2:           NO 2023-11-01 08:31:00            20.9              47.2
+#>    3:           NO 2023-11-01 08:32:00            21.0              46.8
+#>    4:           NO 2023-11-01 08:33:00            20.8              47.4
+#>    5:           NO 2023-11-01 08:34:00            20.6              48.2
 #>   ---                                                                   
-#> 1433:           NO 2023-08-29 19:26:00            12.7              77.1
-#> 1434:           NO 2023-08-29 19:27:00            12.7              77.2
-#> 1435:           NO 2023-08-29 19:28:00            12.6              77.2
-#> 1436:           NO 2023-08-29 19:29:00            12.5              77.3
-#> 1437:           NO 2023-08-29 19:30:00            12.5              77.5
-#>       soil_temperature solar_irradiance rainfall dew_point wet_bulb
-#>    1:             20.7                1        0      10.7     13.7
-#>    2:             20.7                1        0      10.7     13.7
-#>    3:             20.6                1        0      10.6     13.6
-#>    4:             20.6                1        0      10.7     13.6
-#>    5:             20.6                1        0      10.7     13.6
-#>   ---                                                              
-#> 1433:             19.6                1        0       8.8     10.9
-#> 1434:             19.5                1        0       8.8     10.9
-#> 1435:             19.5                1        0       8.7     10.8
-#> 1436:             19.5                1        0       8.6     10.7
-#> 1437:             19.5                1        0       8.7     10.7
-#>       wind_height wind_avg_speed wind_avg_direction_compass_point
-#>    1:           3           3.17                               SW
-#>    2:           3           2.66                              SSW
-#>    3:           3           3.17                                S
-#>    4:           3           1.66                              SSW
-#>    5:           3           1.15                               SW
-#>   ---                                                            
-#> 1433:           3           2.16                               SE
-#> 1434:           3           2.16                               SE
-#> 1435:           3           1.15                               SE
-#> 1436:           3           2.16                               SE
-#> 1437:           3           3.17                              SSE
-#>       wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>    1:                        224          2.160           4.18
-#>    2:                        203          1.656           4.68
-#>    3:                        172          2.160           4.68
-#>    4:                        213          0.144           5.18
-#>    5:                        236          0.648           3.17
[…]

[…]
