+++ R/get_dpird_summaries.R
@@ -187,7 +187,12 @@
 get_dpird_summaries <- function(station_code,
                                 start_date,
                                 end_date = Sys.Date(),
-                                interval = "daily",
+                                interval = c("daily",
+                                             "15min",
+                                             "30min",
+                                             "hourly",
+                                             "monthly",
+                                             "yearly"),
                                 values = "all",
                                 include_closed = FALSE,
                                 api_key) {
