+++ tests/testthat/test-get_patched_point.R
@@ -45,7 +45,7 @@ test_that("get_patched_point() returns daily values", {
         station_code = "008137",
         start_date = "2021-06-01",
         end_date = "2021-07-01",
-        api_key = Sys.getenv("SILO_API_KEY"),
+        api_key = Sys.getenv("SILO_API_KEY")
       )
     )
   })
@@ -116,7 +116,7 @@ test_that("get_patched_point() returns selected daily values", {
                           values = c("rain", "max_temp", "min_temp"),
                           start_date = "2021-06-01",
                           end_date = "2021-07-01",
-                          api_key = Sys.getenv("SILO_API_KEY"),
+                          api_key = Sys.getenv("SILO_API_KEY")
                         )
                       )
                     })
