+++ tests/testthat/test-get_data_drill.R
@@ -56,7 +56,7 @@ test_that("get_data_drill() returns all daily values", {
       longitude = 150.05,
       start_date = "2021-06-01",
       end_date = "2021-07-01",
-      api_key = Sys.getenv("SILO_API_KEY"),
+      api_key = Sys.getenv("SILO_API_KEY")
     )
   })
   expect_equal(nrow(wd), 31)
@@ -123,7 +123,7 @@ test_that("get_data_drill() returns selected daily values", {
       values = c("rain", "max_temp", "min_temp"),
       start_date = "2021-06-01",
       end_date = "2021-07-01",
-      api_key = Sys.getenv("SILO_API_KEY"),
+      api_key = Sys.getenv("SILO_API_KEY")
     )
   })
   expect_equal(nrow(wd), 31)
