### 2023-11-03 1266e79
Revert "Replace e-mail address with Renviron object"
This reverts commit 2b7be87a28d2810001d93d30a41f1678b6f2554c.
