### 2023-11-03 d53cc8a
Fixes different date types depending on the API that's queried (#44)
