+++ vignettes/weatherOz_for_BOM.Rmd
@@ -71,59 +71,59 @@ library(weatherOz)
 #>   4:     3   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
 #>   5:     4   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
 #>  ---                                                                      
-#> 787:     2   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 788:     3   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 789:     4   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 790:     5   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
-#> 791:     6   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 900:     3   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 901:     4   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 902:     5   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 903:     6   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
+#> 904:     7   IDQ11295   QLD Port Douglas QLD_PT254 -16.48681 145.4635 70.4
 #>         start_time_local end_time_local utc_offset      start_time_utc
-#>   1: 2023-11-02 08:49:16     2023-11-03      10:00 2023-11-01 22:49:16
-#>   2: 2023-11-03 00:00:00     2023-11-04      10:00 2023-11-02 14:00:00
-#>   3: 2023-11-04 00:00:00     2023-11-05      10:00 2023-11-03 14:00:00
-#>   4: 2023-11-05 00:00:00     2023-11-06      10:00 2023-11-04 14:00:00
-#>   5: 2023-11-06 00:00:00     2023-11-07      10:00 2023-11-05 14:00:00
+#>   1: 2023-11-03 17:00:00     2023-11-04      10:00 2023-11-03 07:00:00
+#>   2: 2023-11-04 00:00:00     2023-11-05      10:00 2023-11-03 14:00:00
+#>   3: 2023-11-05 00:00:00     2023-11-06      10:00 2023-11-04 14:00:00
+#>   4: 2023-11-06 00:00:00     2023-11-07      10:00 2023-11-05 14:00:00
+#>   5: 2023-11-07 00:00:00     2023-11-08      10:00 2023-11-06 14:00:00
 #>  ---                                                                  
-#> 787: 2023-11-04 00:00:00     2023-11-05      10:00 2023-11-03 14:00:00
-#> 788: 2023-11-05 00:00:00     2023-11-06      10:00 2023-11-04 14:00:00
-#> 789: 2023-11-06 00:00:00     2023-11-07      10:00 2023-11-05 14:00:00
-#> 790: 2023-11-07 00:00:00     2023-11-08      10:00 2023-11-06 14:00:00
-#> 791: 2023-11-08 00:00:00     2023-11-09      10:00 2023-11-07 14:00:00
+#> 900: 2023-11-06 00:00:00     2023-11-07      10:00 2023-11-05 14:00:00
+#> 901: 2023-11-07 00:00:00     2023-11-08      10:00 2023-11-06 14:00:00
+#> 902: 2023-11-08 00:00:00     2023-11-09      10:00 2023-11-07 14:00:00
[…]
