+++ vignettes/weatherOz_for_SILO.Rmd
@@ -105,18 +105,18 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude year month  day       date air_tmax air_tmax_source air_tmin
-#>   1:       150      -27 2022  2022 2022 2022-01-01     29.3              25     18.8
-#>   2:       150      -27 2022  2022 2022 2022-01-02     32.4              25     16.8
-#>   3:       150      -27 2022  2022 2022 2022-01-03     34.6              25     18.2
-#>   4:       150      -27 2022  2022 2022 2022-01-04     35.0              25     18.0
-#>   5:       150      -27 2022  2022 2022 2022-01-05     35.5              25     20.4
-#>  ---                                                                                
-#> 666:       150      -27 2023  2023 2023 2023-10-28     27.8              25      9.5
-#> 667:       150      -27 2023  2023 2023 2023-10-29     28.8              25     11.5
-#> 668:       150      -27 2023  2023 2023 2023-10-30     33.2              25     14.0
-#> 669:       150      -27 2023  2023 2023 2023-10-31     38.5              25     18.3
-#> 670:       150      -27 2023  2023 2023 2023-11-01     30.5              75     16.4
+#>      longitude latitude year month day       date air_tmax air_tmax_source air_tmin
+#>   1:       150      -27 2022     1   1 2022-01-01     29.3              25     18.8
+#>   2:       150      -27 2022     1   2 2022-01-02     32.4              25     16.8
+#>   3:       150      -27 2022     1   3 2022-01-03     34.6              25     18.2
+#>   4:       150      -27 2022     1   4 2022-01-04     35.0              25     18.0
+#>   5:       150      -27 2022     1   5 2022-01-05     35.5              25     20.4
+#>  ---                                                                               
+#> 667:       150      -27 2023    10  29 2023-10-29     28.8              25     11.5
+#> 668:       150      -27 2023    10  30 2023-10-30     33.2              25     14.0
+#> 669:       150      -27 2023    10  31 2023-10-31     38.5              25     18.3
+#> 670:       150      -27 2023    11   1 2023-11-01     32.9              25     16.4
+#> 671:       150      -27 2023    11   2 2023-11-02     30.6              75     14.6
 #>      air_tmin_source  elev_m et_morton_actual et_morton_actual_source
 #>   1:              25 274.9 m              6.5                      26
 #>   2:              25 274.9 m              6.9                      26
@@ -124,11 +124,11 @@ library(weatherOz)
 #>   4:              25 274.9 m              6.1                      26
 #>   5:              25 274.9 m              6.4                      26
 #>  ---                                                                 
-#> 666:              25 274.9 m              3.4                      26
 #> 667:              25 274.9 m              3.5                      26
 #> 668:              25 274.9 m              3.6                      26
 #> 669:              25 274.9 m              0.8                      26
 #> 670:              25 274.9 m              0.0                      26
+#> 671:              25 274.9 m              2.8                      26
 #>      et_morton_potential et_morton_potential_source et_morton_wet
 #>   1:                 7.2                         26           6.8
 #>   2:                 8.9                         26           7.9
@@ -136,11 +136,11 @@ library(weatherOz)
 #>   4:                10.2                         26           8.1
 #>   5:                10.4                         26           8.4
 #>  ---                                                             
-#> 666:                 8.3                         26           5.8
 #> 667:                 8.7                         26           6.1
 #> 668:                 9.7                         26           6.6
[…]

[…]
