### 2023-11-26 db09382
Add checks for earliest available BOM data. This is already done for DPIRD data
