### 2023-11-26 8023056
Add documentation to internal functions
