### 2023-11-26 54790d9
Add checks for earliest available BOM data. (#48)
