+++ R/get_data_drill_apsim.R
@@ -126,6 +126,7 @@ get_data_drill_apsim <- function(longitude,
   start_date <- .check_date(start_date)
   end_date <- .check_date(end_date)
   .check_date_order(start_date, end_date)
+  .check_earliest_available(start_date)
 
   # reformat date for sending to SILO
   start_date <- gsub("-", "", start_date)
