### 2023-11-26 fcfbe18
Ensure that all internal functions are properly tagged
