### 2023-11-27 8c04814
round lonlat to 4 digits
