### 2023-11-27 eff0d36
Fix checks for earliest start dates for respective APIs
