+++ R/get_patched_point.R
@@ -197,7 +197,7 @@ get_patched_point <- function(station_code,
   start_date <- .check_date(start_date)
   end_date <- .check_date(end_date)
   .check_date_order(start_date, end_date)
-  .check_earliest_available(start_date)
+  .check_earliest_available_silo(start_date)
 
   # reformat date for sending to SILO
   start_date <- gsub("-", "", start_date)
