+++ R/get_dpird_summaries.R
@@ -10,10 +10,10 @@
 #'
 #' # Column Name Details
 #'
-#' Column names are converted from the default returns of the API to be
-#'    snake_case formatted and where appropriate, the names of the values that
-#'    are analogous between \acronym{SILO} and \acronym{DPIRD} data are named
-#'    using the same name for ease of interoperability, _e.g._, using
+#' Column names are converted from the default returns of the \acronym{API} to
+#'    be snake_case formatted and where appropriate, the names of the values
+#'    that are analogous between \acronym{SILO} and \acronym{DPIRD} data are
+#'    named using the same name for ease of interoperability, _e.g._, using
 #'    `rbind()` to create a `data.table` that contains data from both APIs.
 #'    However, use with caution and don't mix datasets of different time-steps,
 #'    _i.e._, this function gets many summary values not just \dQuote{daily}
