+++ man/get_dpird_summaries.Rd
@@ -79,10 +79,10 @@ The earliest available data start from August of 2000 for Vasse, \dQuote{VA}.
 }
 
 \section{Column Name Details}{
-Column names are converted from the default returns of the API to be
-snake_case formatted and where appropriate, the names of the values that
-are analogous between \acronym{SILO} and \acronym{DPIRD} data are named
-using the same name for ease of interoperability, \emph{e.g.}, using
+Column names are converted from the default returns of the \acronym{API} to
+be snake_case formatted and where appropriate, the names of the values
+that are analogous between \acronym{SILO} and \acronym{DPIRD} data are
+named using the same name for ease of interoperability, \emph{e.g.}, using
 \code{rbind()} to create a \code{data.table} that contains data from both APIs.
 However, use with caution and don't mix datasets of different time-steps,
 \emph{i.e.}, this function gets many summary values not just \dQuote{daily}
