### 2023-11-30 #598 @rogerssam
HI all, here is my review.

## Package Review

    -   No working relationship with authors to date.
-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☒ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☒ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 6

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

Overall the package appears to be well developed, robust and meets a need of providing access to weather data in Australia. Some additional comments:

-   Checking for the example API key is a nice UX touch.
-   The vignettes are clear, and show a good demonstration of the functionality of the package.
-   The main functions follow a clear convention, which provides obvious entry points for new users.
-   Great to see near complete test coverage.
-   I am familiar with the difference between patched point and data drills from the SILO data, but I expect new users would not be, so perhaps the titles of those functions could be expanded? For example `Get DataDrill (Gridded) Weather Data From SILO` and `Get PatchedPoint (Station) Weather Data From SILO` or similar. I think it makes sense to have the function names matching the SILO API, but I think the titles could provide a little more clarity. I also realise that the detail is available within the function documentation, however that’s an extra step for users to find the right function that they need.

In the current state, I think the package is very good and perfectly usable, however below are some additional suggestions for improvements. I will add some issues on the repo for consideration by the authors.

-   I wonder if it’s worth including some sort of validation on the `api_key` (email) input to the SILO functions? Obviously [email address validation is not trivial](https://stackoverflow.com/questions/201323/how-can-i-validate-an-email-address-using-a-regular-expression) but at present entry of `abc` as the api\_key/email is valid input. Some obvious errors are caught by SILO (e.g. `---@...`) but others are not. I will leave it as a decision for the authors.
-   I think it would be useful for the `get_station_metadata()` function to have the ability to return a single station metadata, rather than just all of them. There could be a `station` argument set to `"all"` by default.

Overall, great work, and I will be personally using it shortly for some of my work!

Thanks,
Sam Rogers
