### 2023-11-30 DPIRD-FSI/weatherOz#49 @rogerssam
SILO API key (email) not validated
Hi all,

As per my [rOpenSci review](https://github.com/ropensci/software-review/issues/598#issuecomment-1832968340), I wonder if it’s worth including some sort of validation on the api_key (email) input to the SILO functions?

Obviously [email address validation is not trivial](https://stackoverflow.com/questions/201323/how-can-i-validate-an-email-address-using-a-regular-expression) but at present entry of `abc` as the api_key/email is valid input. Some obvious errors are caught by SILO (e.g. ---@...) but others are not. I don't think this needs complete email validation, but perhaps some checks would be worthwhile (mostly to reduce the likelihood of this package being the source of API spamming).

Reprex for example:

```r
# works
get_patched_point("008137", "2022-01-01", api_key = "abc")
get_patched_point("008137", "2022-01-01", api_key = "a@b")

# fails
get_patched_point("23031", "2022-01-01", api_key = "---@...")
```
