### 2023-11-30 DPIRD-FSI/weatherOz#50 @rogerssam
FR: Return single station metadata
As per my [rOpenSci review](https://github.com/ropensci/software-review/issues/598#issuecomment-1832968340) I think it would be useful for the get_station_metadata() function to have the ability to return a single station metadata, rather than just all of them. There could be a station argument set to "all" by default.

Current behaviour:

```r
get_station_metadata(which_api = "silo", api_key="abc")
```
Returns:
```
station_code                     station_name      start        end latitude longitude
   1:       061065                Aberdeen Rossgole 1926-01-01 2023-11-30 -32.1402  150.7285
   2:       071000    Adaminaby Alpine Tourist Park 1886-01-01 2023-11-30 -35.9962  148.7693
   3:       068000          Albion Park Post Office 1892-01-01 2023-11-30 -34.5712  150.7761
   4:       068241 Albion Park Shellharbour Airport 1999-01-01 2023-11-30 -34.5639  150.7924
   5:       072146                   Albury Airport 1973-01-01 2023-11-30 -36.0694  146.9531
  ---
3718:       007096                             Yuin 1898-01-01 2023-11-30 -27.9808  116.0347
3719:       012092                         Yuinmery 1921-01-01 2023-11-30 -28.5606  119.0161
3720:       002030                           Yulmbu 1950-01-01 2023-11-30 -17.2986  126.9158
3721:       008147                             Yuna 1909-01-01 2023-11-30 -28.3250  114.9589
3722:       012093                      Yundamindra 1901-01-01 2023-11-30 -29.2472  122.1022
      state elev_m                source status   wmo
   1:   NSW    543 Bureau of Meteorology   open    NA
   2:   NSW   1015 Bureau of Meteorology   open    NA
   3:   NSW      8 Bureau of Meteorology   open    NA
   4:   NSW      8 Bureau of Meteorology   open 95748
   5:   NSW    165 Bureau of Meteorology   open    NA
  ---
3718:    WA    300 Bureau of Meteorology   open    NA
3719:    WA    500 Bureau of Meteorology   open    NA
3720:    WA    450 Bureau of Meteorology   open    NA
3721:    WA    270 Bureau of Meteorology   open    NA
3722:    WA    450 Bureau of Meteorology   open    NA
```

But ideally, behaviour would be:

```r
get_station_metadata(station = "061065", which_api = "silo", api_key="abc")

```

returns:

```
station_code                     station_name      start        end latitude longitude
   1:       061065                Aberdeen Rossgole 1926-01-01 2023-11-30 -32.1402  150.7285
 state elev_m                source status   wmo
   1:   NSW    543 Bureau of Meteorology   open    NA
```
