### 2023-11-30 DPIRD-FSI/weatherOz#50 @adamhsparks
Could do. Easy enough to just subset the station(s) of interest and return only those if requested.
