### 2023-11-30 DPIRD-FSI/weatherOz#50 @rogerssam
Yeah fair call. I figured it was probably more efficient in only requesting a single station if so desired, but it's probably not a big difference either way.
