### 2023-11-30 DPIRD-FSI/weatherOz#50 @adamhsparks
It's possible with the DPIRD API to request a single station, but how it's set up with SILO, we still fetch the entire list from BOM first to get data that SILO doesn't provide.
