### 2023-12-02 dfb006c
Adds functionality to filter metadata results by station code or name.
