### 2023-12-02 DPIRD-FSI/weatherOz#50 @adamhsparks
@rogerssam, @bozaah, I've create PR to address this issue.

Please have a look. If it's acceptable, please approve and merge.

https://github.com/DPIRD-FSI/weatherOz/pull/51
