### 2023-12-02 4bfd54c
Tidy up references to data.table
