+++ R/get_patched_point.R
@@ -105,7 +105,7 @@
 #'       observations for that day of year.}
 #'   }
 #'
-#' @return a [data.table::data.table] with the weather data queried with the
+#' @return a [data.table] with the weather data queried with the
 #'   weather variables in alphabetical order.  The first eight columns will
 #'   always be:
 #'
