+++ man/parse_ag_bulletin.Rd
@@ -14,7 +14,7 @@ or postal code.  Fuzzy string matching via \code{\link[base:agrep]{base::agrep()
 file(s) to parse.  See Details for more.}
 }
 \value{
-A \link[data.table:data.table]{data.table::data.table} of Australia \acronym{BOM} agricultural
+A \link{data.table} of Australia \acronym{BOM} agricultural
 bulletin information.
 }
 \description{
