### 2023-12-06 8166c92
Returns a single (or multiple) station in get_station_metadata() (#51)
* Clarity in docs

* Adds functionality to filter metadata results by station code or name.

* Update WORDLIST

* Update codemeta.json

* Fix cross-ref to data.table

* Tidy up references to data.table
