+++ R/parse_precis_forecast.R
@@ -32,7 +32,7 @@
 #'   files are all present in the directory.
 #'
 #' @return
-#' A [data.table::data.table] of Australia \acronym{BOM} précis seven-day
+#' A [data.table] of Australia \acronym{BOM} précis seven-day
 #'   forecasts for \acronym{BOM} selected towns.
 #'
 #' @examplesIf interactive()
