+++ R/find_forecast_towns.R
@@ -16,8 +16,8 @@
 #' # find forecast towns near Esperance, WA
 #' find_forecast_towns(longitude = 121.8913, latitude = -33.8614)
 #'
-#' @return A [data.table::data.table] of all forecast towns (in this package)
-#'   sorted by distance from \var{latitude} and \var{longitude}, ascending.
+#' @return A [data.table] of all forecast towns (in this package) sorted by
+#'   distance from \var{latitude} and \var{longitude}, ascending.
 #' @author Hugh Parsonage, \email{hugh.parsonage@@gmail.com}, and James Goldie,
 #'   \email{me@@jamesgoldie.dev}, and Adam H. Sparks,
 #'   \email{adam.sparks@@dpird.wa.gov.au}
