### 2023-12-11 3dfba0b
Closes 49: add email validation for SILO API requests
