### 2023-12-11 DPIRD-FSI/weatherOz#49 @bozaah
Thank you @rogerssam. I have added an email validation check ([3dfba0b](https://github.com/DPIRD-FSI/weatherOz/commit/3dfba0b0720db7bb5738afa9d1169b07d8abdbc0)) which should be in line with your suggestion.
