### 2023-12-11 #598 @bozaah
Dear @ldecicco-USGS,
We have addressed the feedback from @rogerssam which we appreciate a lot. We also took on board the comments regarding the name of functions requesting SILO data and the documentation on the data types accepted as `station_code` by the SILO functions. We will address those once we receive the review from @JanLauGe if that is acceptable. Much appreciated, Rodrigo.
