### 2023-12-20 b438d25
testing updated tic.R with pak installation and cache
