+++ .github/workflows/tic.yml
@@ -47,7 +47,7 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
 
       - uses: r-lib/actions/setup-r@v2
         id: install-r
@@ -66,23 +66,8 @@ jobs:
           saveRDS(pak::pkg_deps("local::.", dependencies = TRUE), ".github/r-depends.rds")
         shell: Rscript {0}
 
-      - name: "[Cache] Restore R package cache"
-        uses: actions/cache@v2
-        with:
-          path: |
-            ${{ env.R_LIBS_USER }}/*
-            !${{ env.R_LIBS_USER }}/pak
-          key: ${{ matrix.config.os }}-${{ steps.install-r.outputs.installed-r-version }}-1-${{ hashFiles('.github/r-depends.rds') }}
-          restore-keys: ${{ matrix.config.os }}-${{ steps.install-r.outputs.installed-r-version }}-1-
-
       - name: "[Stage] Install system dependencies"
         if: runner.os == 'Linux'
-        run: |
-          pak::local_system_requirements(execute = TRUE)
-          pak::pkg_system_requirements("rcmdcheck", execute = TRUE)
-        shell: Rscript {0}
-
-      - name: "[Stage] Install dependencies"
         run: |
           pak::local_install_dev_deps(upgrade = TRUE)
           pak::pkg_install("rcmdcheck")
@@ -91,26 +76,29 @@ jobs:
       # set date/week for use in cache creation
       # https://github.community/t5/GitHub-Actions/How-to-set-and-access-a-Workflow-variable/m-p/42970
       # - cache R packages daily
-      # - name: "[Cache] Prepare daily timestamp for cache"
-      #  if: runner.os != 'Windows'
-      #  id: date
-      #  run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
+      - name: "[Cache] Prepare daily timestamp for cache"
+        if: runner.os != 'Windows'
+        id: date
+        run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
 
-      # - name: "[Cache] Cache R packages"
-      #  if: runner.os != 'Windows'
-      #  uses: pat-s/always-upload-cache@v3
-      #  with:
-      #    path: ${{ env.R_LIBS_USER }}
-      #    key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-      #    restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-
-      # - name: "[Stage] Install pak"
-      #  run: Rscript -e "install.packages('pak', repos = 'https://r-lib.github.io/p/pak/stable')"
-
-      # "if (grepl('Ubuntu', Sys.info()[['version']]) && !grepl('Under development', R.version[['status']])) {options(repos = c(CRAN = sprintf('https://packagemanager.rstudio.com/all/__linux__/%s/latest', system('lsb_release -cs', intern = TRUE))))}else{options(repos = c(CRAN = 'cloud.r-project.org'))}; pak::pkg_install('ropensci/tic')"
+      - name: "[Cache] Restore R package cache"
+        if: runner.os != 'Windows'
+        uses: actions/cache/restore@v3
+        with:
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
 
       - name: "[Stage] Install"
[…]
