### 2023-12-20 aa0d7a9
split tic installation from script
