+++ .github/workflows/tic.yml
@@ -88,14 +88,17 @@ jobs:
           path: ${{ env.R_LIBS_USER }}
           key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
 
-      - name: "[Stage] Install"
-        run: Rscript -e "install.packages('tic', repos = c('https://ropensci.r-universe.dev', if (grepl('Ubuntu', Sys.info()[['version']])) {sprintf('https://packagemanager.rstudio.com/all/__linux__/%s/latest', system('lsb_release -cs', intern = TRUE))} else {'https://cloud.r-project.org'}))" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
+      - name: "[Stage] Install tic"
+        run: Rscript -e "pak::pkg_install('ropensci/tic', dependencies = TRUE)"
 
-      - name: "[Cache] Save R package cache"
-        if: runner.os != 'Windows' && always()
-        uses: actions/cache/save@v3
-        with:
-          path: |
+      - name: "[Stage] Install"
+        run: Rscript -e "if (grepl('Ubuntu', Sys.info()[['version']]) && !grepl('Under development', R.version[['status']])) {options(repos = c(CRAN = sprintf('https://packagemanager.rstudio.com/all/__linux__/%s/latest', system('lsb_release -cs', intern = TRUE))))}else{options(repos = c(CRAN = 'cloud.r-project.org'))}" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
+
+        - name: "[Cache] Save R package cache"
+          if: runner.os != 'Windows' && always()
+          uses: actions/cache/save@v3
+          with:
+           path: |
             ${{ env.R_LIBS_USER }}/*
             !${{ env.R_LIBS_USER }}/pak
           key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
@@ -112,6 +115,7 @@ jobs:
         with:
           name: ${{ runner.os }}-r${{ matrix.config.r }}-results
           path: check
+
       - name: "[Stage] Before Deploy"
         run: |
           Rscript -e "tic::before_deploy()"
