+++ .github/workflows/tic.yml
@@ -92,13 +92,13 @@ jobs:
         run: Rscript -e "pak::pkg_install('ropensci/tic', dependencies = TRUE)"
 
       - name: "[Stage] Install"
-        run: Rscript -e "if (grepl('Ubuntu', Sys.info()[['version']]) && !grepl('Under development', R.version[['status']])) {options(repos = c(CRAN = sprintf('https://packagemanager.rstudio.com/all/__linux__/%s/latest', system('lsb_release -cs', intern = TRUE))))}else{options(repos = c(CRAN = 'cloud.r-project.org'))}" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
+        run: Rscript -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
 
-        - name: "[Cache] Save R package cache"
-          if: runner.os != 'Windows' && always()
-          uses: actions/cache/save@v3
-          with:
-           path: |
+      - name: "[Cache] Save R package cache"
+        if: runner.os != 'Windows' && always()
+        uses: actions/cache/save@v3
+        with:
+          path: |
             ${{ env.R_LIBS_USER }}/*
             !${{ env.R_LIBS_USER }}/pak
           key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
