### 2023-12-21 #598 @JanLauGe
No working relationship with authors to date.

I have not tested all functions yet due to getting stuck with the unit tests.
Currently I'm getting test errors on my local machine. Happy to work with you to figure out what is going on.

-----------------

### Additional comments:

* [this](https://github.com/DPIRD-FSI/weatherOz/blob/3dfba0b0720db7bb5738afa9d1169b07d8abdbc0/R/get_station_metadata.R#L123) check for whether the API key has been provided currently doesn't work with the examples because both `NULL` and an empty string resulting from `Sys.getenv("DPIRD_API_KEY")` passes the test. That means that if a new user just uses the function from the example they end up with a relatively cryptic error. Can you add a test or conditition for `NULL` and empty string?

* I agree with @rogerssam that email address validation would be a good addition. To play it safe, the package could produce a `message()` when detecting an invalid email address instead of `stop()`-ing with an error. [Example](https://www.r-bloggers.com/2012/07/validating-email-adresses-in-r/) of an email address validation regex.

* Running tests required installing the `vcr` package for me which is not listed as dependency.

* Currently some tests are failing for me locally (iOS). I'm attaching the test report below.

```
...
trying URL 'ftp://ftp.bom.gov.au/anon/gen/fwo/IDW14199.xml'
Content type 'unknown' length 546501 bytes (533 KB)
==================================================
file downloaded to:/var/folders/43/c5m9cb696l7fgnfsls66r1f400th2k/T//RtmpbdumdW/file6cd638754f36.gif
file downloaded to:/var/folders/43/c5m9cb696l7fgnfsls66r1f400th2k/T//RtmpbdumdW/file6cd61da37f2f.gif

── Error ('test-get_dpird_summaries.R:235:15'): get_dpird_summaries() returns hourly values ──

================================================================================
An HTTP request has been made that vcr does not know how to handle:
GET https://api.dpird.wa.gov.au/v2/weather/stations/summaries/hourly?stationCode=BI&startDateTime=2017-10-28&endDateTime=2017-11-01&interval=hourly&select=stationCode%2CstationName%2Cperiod%2Cwind&group=all&includeClosed=false&api_key=<<<dpird_api_key>>>&offset=0&limit=74
vcr is currently using the following cassette:
  - <local path redacted>/weatherOz/tests/fixtures/dpird_summaries_hourly.json
    - record_mode: once
    - match_requests_on: method, uri
```
