### 2023-12-22 #598 @yabellini
Hi @rogerssam, @arhepworth, @syntrix, @becoleary. rOpenSci Community Manager here.
If you want to join our Slack, I can send an invitation.  Please share with me what email you want to use.  You can write me to yabellini@ropensci.org.  Thank you!
