### 2024-01-03 #598 @ldecicco-USGS
Thanks so much @JanLauGe! If you aren't on the rOpenSci slack, please see @yabellini 's message above
