### 2024-01-04 051f5a1
fix issue with 15 and 30 min dpird summaries, continue WIP with ghactions/tic
