+++ .github/workflows/main.yml
@@ -1,5 +1,5 @@
 ## tic GitHub Actions template: linux-macos-windows-deploy
-## revision date: 2023-12-20
+## revision date: 2024-01-04
 on:
   workflow_dispatch:
   push:
@@ -7,7 +7,6 @@ on:
     - main
     - master
     - cran-*
-    - develop
   pull_request:
     branches:
     - main
@@ -45,69 +44,50 @@ jobs:
       BUILD_PKGDOWN: ${{ matrix.config.pkgdown }}
       # use GITHUB_TOKEN from GitHub to workaround rate limits in {remotes}
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_KEEP_PKG_SOURCE: yes
 
     steps:
       - uses: actions/checkout@v4
 
       - uses: r-lib/actions/setup-r@v2
         with:
+          rtools-version: '42'
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
-          Ncpus: 4
 
       - uses: r-lib/actions/setup-tinytex@v2
         if: matrix.config.latex == 'true'
 
       - uses: r-lib/actions/setup-pandoc@v2
 
-      - name: Install macOS system dependencies
-        if: runner.os == 'macOS'
-        run: brew install gdal proj
-
-      - name: Install pak and query dependencies
-        run: |
-          install.packages("pak", repos = "https://r-lib.github.io/p/pak/dev/")
-          saveRDS(pak::pkg_deps("local::.", dependencies = TRUE), ".github/r-depends.rds")
-        shell: Rscript {0}
-
-      #- name: "[Stage] Install system dependencies"
-      #  if: runner.os == 'Linux'
-      #  run: |
-      #    pak::local_install_dev_deps(upgrade = TRUE)
-      #    pak::pkg_install("rcmdcheck")
-      #  shell: Rscript {0}
+      - name: Install sys deps for Ubuntu
+        if: runner.os == 'Linux'
+        run: sudo apt update && sudo apt install -y libgit2-dev libcurl4-openssl-dev
 
       # set date/week for use in cache creation
       # https://github.community/t5/GitHub-Actions/How-to-set-and-access-a-Workflow-variable/m-p/42970
       # - cache R packages daily
-      #- name: "[Cache] Prepare daily timestamp for cache"
-      #  if: runner.os != 'Windows'
-      #  id: date
-      #  run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
+      - name: "[Cache] Prepare daily timestamp for cache"
+        if: runner.os != 'Windows'
+        id: date
+        run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
 
-      #- name: "[Cache] Restore R package cache"
-      #  if: runner.os != 'Windows'
-      #  uses: actions/cache/restore@v3
-      #  with:
-      #    path: ${{ env.R_LIBS_USER }}
-      #    key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-
-      - name: "[Stage] Install tic"
-        run: Rscript -e "pak::pkg_install('ropensci/tic', dependencies = TRUE)"
+      - name: "[Cache] Restore R package cache"
+        if: runner.os != 'Windows'
+        uses: actions/cache/restore@v3
+        with:
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
 
       - name: "[Stage] Install"
[…]
