### 2024-01-04 30425cd
Revert "fix issue with 15 and 30 min dpird summaries, continue WIP with ghactions/tic"
This reverts commit 051f5a19f39541d8a9907adb80d1bae7bda87c7e.
