### 2024-01-04 8348cb3
fix issue with 15 and 30 min dpird summaries, update tic and rename
