+++ .github/workflows/main.yml
@@ -52,6 +52,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
+          rtools-version: '42'
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
           use-public-rspm: true
@@ -66,48 +67,27 @@ jobs:
         if: runner.os == 'macOS'
         run: brew install gdal proj
 
-      - name: Install pak and query dependencies
-        run: |
-          install.packages("pak", repos = "https://r-lib.github.io/p/pak/dev/")
-          saveRDS(pak::pkg_deps("local::.", dependencies = TRUE), ".github/r-depends.rds")
-        shell: Rscript {0}
+      - name: "[Cache] Prepare daily timestamp for cache"
+        if: runner.os != 'Windows'
+        id: date
+        run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
 
-      #- name: "[Stage] Install system dependencies"
-      #  if: runner.os == 'Linux'
-      #  run: |
-      #    pak::local_install_dev_deps(upgrade = TRUE)
-      #    pak::pkg_install("rcmdcheck")
-      #  shell: Rscript {0}
-
-      # set date/week for use in cache creation
-      # https://github.community/t5/GitHub-Actions/How-to-set-and-access-a-Workflow-variable/m-p/42970
-      # - cache R packages daily
-      #- name: "[Cache] Prepare daily timestamp for cache"
-      #  if: runner.os != 'Windows'
-      #  id: date
-      #  run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
-
-      #- name: "[Cache] Restore R package cache"
-      #  if: runner.os != 'Windows'
-      #  uses: actions/cache/restore@v3
-      #  with:
-      #    path: ${{ env.R_LIBS_USER }}
-      #    key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-
-      - name: "[Stage] Install tic"
-        run: Rscript -e "pak::pkg_install('ropensci/tic', dependencies = TRUE)"
+      - name: "[Cache] Restore R package cache"
+        if: runner.os != 'Windows'
+        uses: actions/cache/restore@v3
+        with:
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
 
       - name: "[Stage] Install"
-        run: Rscript -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
+        run: Rscript -e "install.packages('tic', repos = c('https://ropensci.r-universe.dev', if (grepl('Ubuntu', Sys.info()[['version']])) {sprintf('https://packagemanager.rstudio.com/all/__linux__/%s/latest', system('lsb_release -cs', intern = TRUE))} else {'https://cloud.r-project.org'}))" -e "print(tic::dsl_load())" -e "tic::prepare_all_stages()" -e "tic::before_install()" -e "tic::install()"
 
-      #- name: "[Cache] Save R package cache"
-      #  if: runner.os != 'Windows' && always()
-      #  uses: actions/cache/save@v3
-      #  with:
-      #    path: |
-      #      ${{ env.R_LIBS_USER }}/*
-      #      !${{ env.R_LIBS_USER }}/pak
-      #    key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
+      - name: "[Cache] Save R package cache"
+        if: runner.os != 'Windows' && always()
[…]
