### 2024-01-04 5509840
added linux deps step to tic
