### 2024-01-04 0f0bd65
added linux deps step to tic workflowc
