+++ .github/workflows/main.yml
@@ -62,124 +62,10 @@ jobs:
         if: matrix.config.latex == 'true'
 
       - uses: r-lib/actions/setup-pandoc@v2
-      ## tic GitHub Actions template: linux-macos-windows-deploy
-## revision date: 2023-12-20
-on:
-  workflow_dispatch:
-  push:
-    branches:
-    - main
-    - master
-    - cran-*
-    - develop
-  pull_request:
-    branches:
-    - main
-    - master
-  # for now, CRON jobs only run on the default branch of the repo (i.e. usually on master)
-  schedule:
-    # * is a special character in YAML so you have to quote this string
-    - cron: "0 4 * * *"
-
-name: tic
-
-jobs:
-  all:
-    runs-on: ${{ matrix.config.os }}
-
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          # use a different tic template type if you do not want to build on all listed platforms
-          - { os: windows-latest, r: "release" }
-          - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
-          - { os: ubuntu-latest, r: "devel", http-user-agent: 'release' }
-          - { os: ubuntu-latest, r: "release" }
-
-    env:
-      # make sure to run `tic::use_ghactions_deploy()` to set up deployment
-      TIC_DEPLOY_KEY: ${{ secrets.TIC_DEPLOY_KEY }}
-      # prevent rgl issues because no X11 display is available
-      RGL_USE_NULL: true
-      # if you use bookdown or blogdown, replace "PKGDOWN" by the respective
-      # capitalized term. This also might need to be done in tic.R
-      BUILD_PKGDOWN: ${{ matrix.config.pkgdown }}
-      # use GITHUB_TOKEN from GitHub to workaround rate limits in {remotes}
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      R_KEEP_PKG_SOURCE: yes
-
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          rtools-version: '42'
-          r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
-          Ncpus: 4
-
-      - uses: r-lib/actions/setup-tinytex@v2
-        if: matrix.config.latex == 'true'
-
-      - uses: r-lib/actions/setup-pandoc@v2
 
       - name: Install sys deps for Ubuntu
         if: runner.os == 'Linux'
         run: sudo apt update && sudo apt install -y libgit2-dev libcurl4-openssl-dev
-
-      - name: Install macOS system dependencies
-        if: runner.os == 'macOS'
-        run: brew install gdal proj
-
-      - name: "[Cache] Prepare daily timestamp for cache"
-        if: runner.os != 'Windows'
-        id: date
-        run: echo "date=$(date '+%d-%m')" >> $GITHUB_OUTPUT
-
-      - name: "[Cache] Restore R package cache"
-        if: runner.os != 'Windows'
-        uses: actions/cache/restore@v3
-        with:
-          path: ${{ env.R_LIBS_USER }}
-          key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
-
-      - name: "[Stage] Install"
[…]
