### 2024-01-10 2b55dcb
fix dpird summary Windows bug
