### 2024-01-10 b1b6416
fix dpird summary Windows bug2
