### 2024-01-30 588df63
Update Adam's e-mail address now that he's no longer at DPIRD
