+++ R/get_coastal_forecast.R
@@ -233,7 +233,7 @@ get_coastal_forecast <- function(state = "AUS") {
 #'
 #' @return a data.table of the forecast for further refining
 #' @keywords internal
-#' @author Adam H. Sparks, \email{adam.sparks@@dpird.wa.gov.au}
+#' @author Adam H. Sparks, \email{adamhsparks@@gmail.com}
 #' @noRd
 
 .parse_coastal_xml <- function(xml_object) {
