+++ R/get_dpird_availability.R
@@ -47,7 +47,7 @@
 #' @family DPIRD
 #' @family metadata
 #'
-#' @author Adam H. Sparks, \email{adam.sparks@@dpird.wa.gov.au}
+#' @author Adam H. Sparks, \email{adamhsparks@@gmail.com}
 #'
 #' @examples
 #' \dontrun{
