+++ R/get_patched_point_apsim.R
@@ -85,7 +85,7 @@
 #'   of Hydrology*, Volume 66, 1-76.
 #'
 #' @author Rodrigo Pires, \email{rodrigo.pires@@dpird.wa.gov.au}, and Adam
-#'   Sparks, \email{adam.sparks@@dpird.wa.gov.au}
+#'   Sparks, \email{adamhsparks@@gmail.com}
 #'
 #' @examples
 #' \dontrun{
