### 2024-02-01 f86b949
add additional checks to api_key args
