+++ R/get_patched_point.R
@@ -181,7 +181,7 @@ get_patched_point <- function(station_code,
          "Please supply a valid start date as `start_date`.")
 
   # Error if api_key is not provided
-  if (missing(api_key)) {
+  if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
     stop(
       "A valid email address must be provided for `api_key`.",
       call. = FALSE
