+++ R/get_data_drill.R
@@ -175,7 +175,7 @@ get_data_drill <- function(longitude,
          "Please supply a valid start date as `start_date`.")
 
   # Error if api_key is not provided
-  if (missing(api_key)) {
+  if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
     stop("A valid email address must be provided for `api_key`.",
          call. = FALSE)
   }
