+++ R/get_patched_point_apsim.R
@@ -121,7 +121,7 @@ get_patched_point_apsim <- function(station_code,
   }
 
   # Error if api_key is not provided
-  if (missing(api_key)) {
+  if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
     stop("A valid email address must be provided for `api_key`.",
          call. = FALSE)
   }
