+++ R/get_dpird_summaries.R
@@ -206,7 +206,7 @@ get_dpird_summaries <- function(station_code,
          "Please supply a valid start date as `start_date`.")
 
   # Error if api_key is not provided
-  if (missing(api_key)) {
+  if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
     stop(
       "A valid DPIRD API key must be provided, please visit\n",
       "<https://www.agric.wa.gov.au/web-apis> to request one.\n",
