+++ R/get_data_drill_apsim.R
@@ -112,7 +112,7 @@ get_data_drill_apsim <- function(longitude,
   }
 
   # Error if api_key is not provided
-  if (missing(api_key)) {
+  if (missing(api_key) | is.null(api_key) | is.na(api_key)) {
     stop("A valid email address must be provided for `api_key`.",
          call. = FALSE)
   }
