### 2024-02-01 ae26fe9
add check to ensure station_code arg is a character string
