### 2024-02-01 #598 @bozaah
Thank you for your time reviewing `{weatherOz}` @JanLauGe and @rogerssam, we appreciate that. Your comments have helped us improve the package and hopefully the user experience with `{weatherOz}`. I have collated the responses to the points you have raised below:

- We have added an email validation check as suggested. Functions will still return an error and `stop` if a valid `api_key` is not provided. Closed issue 49 with [commit 3dfba0b](https://github.com/DPIRD-FSI/weatherOz/commit/3dfba0b0720db7bb5738afa9d1169b07d8abdbc0).

- We have maintained SILO functionality with the same names (i.e., `get_data_drill()` and `get_patched_point()`). As noted, the function documentation is clear on the API endpoints queried and there is also the [SILO vignette](https://dpird-fsi.github.io/weatherOz/articles/weatherOz_for_SILO.html) which details the functionality and has additional information on the endpoints, variables and query output types. We believe this provides the user a clear view of the functionality.

- We have added a check to functions that required the `station_code` argument (i.e., `get_dpird_extremes()`, `get_dpird_minute()`, `get_dpird_summaries()`, `get_patched_point()` and `get_patched_point_apsim()`) so that only a character string type is accepted. [commit ae26fe9](https://github.com/DPIRD-FSI/weatherOz/commit/ae26fe90b5918ea5f3f57975ba5d174ed54cff77).

-  Added functionality to `get_station_metadata()` so that it returns the metadata for only one station if required. Closed issue 51 with [commit 8166c92](https://github.com/DPIRD-FSI/weatherOz/commit/8166c92f63ed138ccfed966a09e9537a35324b67).

- Added check to the functions that require the `api_key` argument (i.e., `get_dpird_*()`, `get_data_drill*()`, `get_patched_point*()`, `find_nearby_stations()` and `get_station_metadata()`) so that missing, empty or null returns an error. [commit f86b949](https://github.com/DPIRD-FSI/weatherOz/commit/f86b9491dec556a3f0e4c7a11ac6d71b18880d96).

- We have listed `{vcr}` in the [Suggests section](https://github.com/DPIRD-FSI/weatherOz/blob/ae26fe90b5918ea5f3f57975ba5d174ed54cff77/DESCRIPTION#L168) of the DESCRIPTION as noted in the [rOpenSci HTTP testing in R book](https://books.ropensci.org/http-testing/vcr.html). We are happy to adjust that if there is better way to address this.

- The test error @JanLauGe is due to the date range used in the query (and consequently saved into the `{vcr}` cassete). The DPIRD endpoint with `hourly` data (and also `15min` and `30min` intervals) only accepts dates up to the 1st day of the  last year (in 2024 this is `2023-01-01`). The function help describes this limitation when the `interval` argument is shorter than `daily`. I have updated the `testthat.R` and saved a new `{vcr}` cassete to reflect that and I have no errors on my current checks/tests. Thank you for picking this up!

I am currently running the checks and tests locally on MacOS and Windows. I have also updated the Github CI/tic workflows and I am hopeful it will work. I have added @JanLauGe and @rogerssam to the list of authors [commit a00dbba](https://github.com/DPIRD-FSI/weatherOz/commit/a00dbba32814559babd235d4684234ed6ba3ec6b) Again, thank you for your feedback, I am sure users will benefit from your comments!

I will create a PR so that @adamhsparks can check for issues I might have missed (or created) while editing `{weatherOz}` in line with your feedback. We will also update the `NEWS.md` with the most recent changes once the merge is accepted.

Please let me know if there is anything else you would like us to take a look or rectify.

Thank you!
