### 2024-02-01 e0defb4
Fixes failing test due to "|" rather than "||" in if()
