### 2024-02-01 9baf26b
Address rOpenSci review process feedback and misc of bug fixes (#52)
Co-authored-by: Adam H. Sparks <adam.sparks@curtin.edu.au>
