### 2024-02-02 #598 @bozaah
I ran the package check before merging the PR! Lets try again,  @ropensci-review-bot check package
