### 2024-02-02 380d9bd
Renames get_stations_metadata and adds roxyglobals
