+++ R/parse_ag_bulletin.R
@@ -64,10 +64,10 @@
 #'
 #' @family BOM
 #' @family parse
-#'
+#' @autoglobal
 #' @seealso [get_ag_bulletin]
 #'
-#' @export parse_ag_bulletin
+#' @export
 #'
 parse_ag_bulletin <- function(state, filepath) {
   the_state <- .check_states(state)
