### 2024-02-02 86c8e1f
Fix bug related to default time-step intervals not being set properly in-function
