+++ R/get_dpird_summaries.R
@@ -238,29 +238,39 @@ get_dpird_summaries <- function(station_code,
   .check_date_order(start_date, end_date)
   .check_earliest_available_dpird(start_date)
 
-  # Use `agrep()` to fuzzy match the user-requested time interval
-  approved_intervals <- c("15min",
-                          "30min",
-                          "hourly",
-                          "daily",
-                          "monthly",
-                          "yearly")
+  # if interval is not set, default to "daily", else check input to be sure
+  if (length(setdiff(interval,
+                     c("daily",
+                       "15min",
+                       "30min",
+                       "hourly",
+                       "monthly",
+                       "yearly"))) == 0) {
+    checked_interval <- "daily"
+  } else {
+    approved_intervals <- c("15min",
+                            "30min",
+                            "hourly",
+                            "daily",
+                            "monthly",
+                            "yearly")
 
-  likely_interval <- agrep(pattern = interval,
-                           x = approved_intervals)
+    likely_interval <- agrep(pattern = interval,
+                             x = approved_intervals)
 
-  # Match time interval query to user requests
-  checked_interval <-
-    try(match.arg(approved_intervals[likely_interval],
-                  approved_intervals,
-                  several.ok = FALSE),
-        silent = TRUE)
+    # Match time interval query to user requests
+    checked_interval <-
+      try(match.arg(approved_intervals[likely_interval],
+                    approved_intervals,
+                    several.ok = FALSE),
+          silent = TRUE)
 
-  # Error if summary interval is not available. API only allows for daily,
-  # 15 min, 30 min, hourly, monthly or yearly
-  if (methods::is(checked_interval, "try-error")) {
-    stop(call. = FALSE,
-         "\"", interval, "\" is not a supported time interval")
+    # Error if summary interval is not available. API only allows for daily,
+    # 15 min, 30 min, hourly, monthly or yearly
+    if (methods::is(checked_interval, "try-error")) {
+      stop(call. = FALSE,
+           "\"", interval, "\" is not a supported time interval")
+    }
   }
 
   request_interval <- lubridate::interval(start_date,
