+++ vignettes/weatherOz_for_SILO.Rmd
@@ -105,114 +105,71 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude year month day       date air_tmax air_tmax_source air_tmin
-#>   1:       150      -27 2022     1   1 2022-01-01     29.3              25     18.8
-#>   2:       150      -27 2022     1   2 2022-01-02     32.4              25     16.8
-#>   3:       150      -27 2022     1   3 2022-01-03     34.6              25     18.2
-#>   4:       150      -27 2022     1   4 2022-01-04     35.0              25     18.0
-#>   5:       150      -27 2022     1   5 2022-01-05     35.5              25     20.4
-#>  ---                                                                               
-#> 667:       150      -27 2023    10  29 2023-10-29     28.8              25     11.5
-#> 668:       150      -27 2023    10  30 2023-10-30     33.2              25     14.0
-#> 669:       150      -27 2023    10  31 2023-10-31     38.5              25     18.3
-#> 670:       150      -27 2023    11   1 2023-11-01     32.9              25     16.4
-#> 671:       150      -27 2023    11   2 2023-11-02     30.6              75     14.6
-#>      air_tmin_source  elev_m et_morton_actual et_morton_actual_source
-#>   1:              25 274.9 m              6.5                      26
-#>   2:              25 274.9 m              6.9                      26
-#>   3:              25 274.9 m              5.3                      26
-#>   4:              25 274.9 m              6.1                      26
-#>   5:              25 274.9 m              6.4                      26
-#>  ---                                                                 
-#> 667:              25 274.9 m              3.5                      26
-#> 668:              25 274.9 m              3.6                      26
-#> 669:              25 274.9 m              0.8                      26
-#> 670:              25 274.9 m              0.0                      26
-#> 671:              25 274.9 m              2.8                      26
-#>      et_morton_potential et_morton_potential_source et_morton_wet
-#>   1:                 7.2                         26           6.8
-#>   2:                 8.9                         26           7.9
-#>   3:                10.7                         26           8.0
-#>   4:                10.2                         26           8.1
-#>   5:                10.4                         26           8.4
-#>  ---                                                             
-#> 667:                 8.7                         26           6.1
-#> 668:                 9.7                         26           6.6
-#> 669:                11.8                         26           6.3
-#> 670:                10.6                         26           5.3
-#> 671:                 7.6                         26           5.2
-#>      et_morton_wet_source et_short_crop et_short_crop_source et_tall_crop
-#>   1:                   26           5.4                   26          6.2
-#>   2:                   26           6.6                   26          7.9
-#>   3:                   26           7.4                   26          9.2
-#>   4:                   26           7.3                   26          9.0
-#>   5:                   26           7.4                   26          9.0
-#>  ---                                                                     
-#> 667:                   26           5.9                   26          7.7
-#> 668:                   26           6.6                   26          8.6
-#> 669:                   26           7.4                   26         10.1
-#> 670:                   26           6.6                   26          9.1
-#> 671:                   26           5.4                   26          6.9
-#>      et_tall_crop_source evap_comb evap_comb_source evap_morton_lake
-#>   1:                  26       4.9               25              7.0
-#>   2:                  26       9.6               25              8.3
-#>   3:                  26       8.2               25              8.4
-#>   4:                  26       8.6               25              8.5
-#>   5:                  26       6.3               25              8.8
-#>  ---                                                                
-#> 667:                  26       7.0               25              6.4
-#> 668:                  26       7.8               25              6.9
-#> 669:                  26       9.0               25              6.5
-#> 670:                  26       6.1               25              5.5
[…]
