+++ vignettes/weatherOz_for_BOM.Rmd
@@ -64,66 +64,45 @@ library(weatherOz)
 #>   Please note information at the page 
 #>   <http://www.bom.gov.au/catalogue/data-feeds.shtml#precis>.
 #>   --------------------------------------------------------------------------------  
-#>      index product_id state         town       aac       lat      lon elev
-#>   1:     0   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
-#>   2:     1   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
-#>   3:     2   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
-#>   4:     3   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
-#>   5:     4   IDQ11295   QLD     Brisbane QLD_PT001 -27.48080 153.0389  8.1
+#>       index product_id  state         town       aac       lat      lon  elev    start_time_local end_time_local utc_offset
+#>      <fctr>     <char> <char>       <char>    <char>     <num>    <num> <num>              <POSc>         <POSc>     <fctr>
+#>   1:      0   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-02-02 08:45:10     2024-02-03      10:00
+#>   2:      1   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-02-03 00:00:00     2024-02-04      10:00
+#>   3:      2   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-02-04 00:00:00     2024-02-05      10:00
+#>   4:      3   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-02-05 00:00:00     2024-02-06      10:00
+#>   5:      4   IDQ11295    QLD     Brisbane QLD_PT001 -27.48080 153.0389   8.1 2024-02-06 00:00:00     2024-02-07      10:00
+#>  ---                                                                                                                       
+#> 787:      2   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-02-04 00:00:00     2024-02-05      10:00
+#> 788:      3   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-02-05 00:00:00     2024-02-06      10:00
+#> 789:      4   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-02-06 00:00:00     2024-02-07      10:00
+#> 790:      5   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-02-07 00:00:00     2024-02-08      10:00
+#> 791:      6   IDQ11295    QLD Port Douglas QLD_PT254 -16.48681 145.4635  70.4 2024-02-08 00:00:00     2024-02-09      10:00
+#>           start_time_utc        end_time_utc minimum_temperature maximum_temperature lower_precipitation_limit
+#>                   <POSc>              <POSc>               <num>               <num>                     <num>
+#>   1: 2024-02-01 22:45:10 2024-02-02 14:00:00                  NA                  33                        NA
+#>   2: 2024-02-02 14:00:00 2024-02-03 14:00:00                  23                  32                        NA
+#>   3: 2024-02-03 14:00:00 2024-02-04 14:00:00                  24                  32                        NA
+#>   4: 2024-02-04 14:00:00 2024-02-05 14:00:00                  23                  32                         0
+#>   5: 2024-02-05 14:00:00 2024-02-06 14:00:00                  23                  33                         0
+#>  ---                                                                                                          
[…]

[…]
