### 2024-02-02 7b883f1
Reknit vignettes and fix any bugs that popped up, e.g. stations with no location
