+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -160,30 +160,30 @@ library(weatherOz)
 #> Key: <station_code>
 #>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall dew_point
 #>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <int>     <num>
-#>    1:           NO 2024-02-01 11:51:00            44.2               9.2             46.1             1096        0       4.6
-#>    2:           NO 2024-02-01 11:52:00            43.6               9.3             46.2             1094        0       4.3
-#>    3:           NO 2024-02-01 11:53:00            43.7               9.3             46.3             1096        0       4.3
-#>    4:           NO 2024-02-01 11:54:00            43.6               9.6             46.4             1096        0       4.7
-#>    5:           NO 2024-02-01 11:55:00            43.6               9.3             46.4             1096        0       4.3
+#>    1:           NO 2024-02-01 12:47:00            44.8               8.8             49.5             1100        0       4.4
+#>    2:           NO 2024-02-01 12:48:00            44.9               8.6             49.6             1101        0       4.1
+#>    3:           NO 2024-02-01 12:49:00            45.1               8.8             49.7             1099        0       4.6
+#>    4:           NO 2024-02-01 12:50:00            45.3               8.5             49.7             1104        0       4.2
+#>    5:           NO 2024-02-01 12:51:00            45.0               8.6             49.8             1100        0       4.2
 #>   ---                                                                                                                        
-#> 1431:           NO 2024-02-02 11:41:00            42.0              10.5             46.4             1081        0       4.8
-#> 1432:           NO 2024-02-02 11:42:00            42.0              10.7             46.5             1083        0       5.1
-#> 1433:           NO 2024-02-02 11:43:00            41.9              10.8             46.5             1084        0       5.1
-#> 1434:           NO 2024-02-02 11:44:00            41.9              10.8             46.6             1085        0       5.1
-#> 1435:           NO 2024-02-02 11:45:00            42.0              10.9             46.6             1084        0       5.3
+#> 1430:           NO 2024-02-02 12:36:00            42.7               9.4             49.4             1105        0       3.8
+#> 1431:           NO 2024-02-02 12:37:00            42.9               9.4             49.5             1103        0       3.9
+#> 1432:           NO 2024-02-02 12:38:00            43.2               9.3             49.6             1101        0       4.0
+#> 1433:           NO 2024-02-02 12:39:00            43.6               8.9             49.6             1104        0       3.6
+#> 1434:           NO 2024-02-02 12:40:00            43.6               8.7             49.7             1102        0       3.3
 #>       wet_bulb wind_height wind_avg_speed wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
 #>          <num>       <int>          <num>                           <char>                      <int>          <num>          <num>
-#>    1:     19.1           3           6.70                               NE                         36          1.656          12.20
[…]
