+++ vignettes/weatherOz_for_BOM.Rmd
@@ -365,42 +365,42 @@ If no valid Product ID is supplied, defaults to all GeoTIFF images currently ava
 library(weatherOz)
 
 (avail <- get_available_imagery(product_id = "IDE00425"))
-#>   [1] "IDE00425.202402010310.tif" "IDE00425.202402010320.tif" "IDE00425.202402010330.tif" "IDE00425.202402010340.tif"
-#>   [5] "IDE00425.202402010350.tif" "IDE00425.202402010400.tif" "IDE00425.202402010410.tif" "IDE00425.202402010420.tif"
-#>   [9] "IDE00425.202402010430.tif" "IDE00425.202402010440.tif" "IDE00425.202402010450.tif" "IDE00425.202402010500.tif"
-#>  [13] "IDE00425.202402010510.tif" "IDE00425.202402010520.tif" "IDE00425.202402010530.tif" "IDE00425.202402010540.tif"
-#>  [17] "IDE00425.202402010550.tif" "IDE00425.202402010600.tif" "IDE00425.202402010610.tif" "IDE00425.202402010620.tif"
-#>  [21] "IDE00425.202402010630.tif" "IDE00425.202402010640.tif" "IDE00425.202402010650.tif" "IDE00425.202402010700.tif"
-#>  [25] "IDE00425.202402010710.tif" "IDE00425.202402010720.tif" "IDE00425.202402010730.tif" "IDE00425.202402010740.tif"
-#>  [29] "IDE00425.202402010750.tif" "IDE00425.202402010800.tif" "IDE00425.202402010810.tif" "IDE00425.202402010820.tif"
-#>  [33] "IDE00425.202402010830.tif" "IDE00425.202402010840.tif" "IDE00425.202402010850.tif" "IDE00425.202402010900.tif"
-#>  [37] "IDE00425.202402010910.tif" "IDE00425.202402010920.tif" "IDE00425.202402010930.tif" "IDE00425.202402010940.tif"
-#>  [41] "IDE00425.202402010950.tif" "IDE00425.202402011000.tif" "IDE00425.202402011010.tif" "IDE00425.202402011020.tif"
-#>  [45] "IDE00425.202402011030.tif" "IDE00425.202402011040.tif" "IDE00425.202402011050.tif" "IDE00425.202402011100.tif"
-#>  [49] "IDE00425.202402011110.tif" "IDE00425.202402011120.tif" "IDE00425.202402011130.tif" "IDE00425.202402011140.tif"
-#>  [53] "IDE00425.202402011150.tif" "IDE00425.202402011200.tif" "IDE00425.202402011210.tif" "IDE00425.202402011220.tif"
-#>  [57] "IDE00425.202402011230.tif" "IDE00425.202402011240.tif" "IDE00425.202402011250.tif" "IDE00425.202402011300.tif"
-#>  [61] "IDE00425.202402011310.tif" "IDE00425.202402011320.tif" "IDE00425.202402011330.tif" "IDE00425.202402011340.tif"
-#>  [65] "IDE00425.202402011350.tif" "IDE00425.202402011400.tif" "IDE00425.202402011410.tif" "IDE00425.202402011420.tif"
-#>  [69] "IDE00425.202402011430.tif" "IDE00425.202402011450.tif" "IDE00425.202402011500.tif" "IDE00425.202402011510.tif"
-#>  [73] "IDE00425.202402011520.tif" "IDE00425.202402011530.tif" "IDE00425.202402011540.tif" "IDE00425.202402011550.tif"
-#>  [77] "IDE00425.202402011600.tif" "IDE00425.202402011610.tif" "IDE00425.202402011620.tif" "IDE00425.202402011630.tif"
[…]
