### 2024-02-02 4749350
update NEWS.md
