+++ NEWS.md
@@ -1,5 +1,9 @@
 # weatherOz 0.0.1.9000
 
+## Major Changes
+
+* Changes the name of the function `get_station_metadata()` to `get_stations_metadata()` to avoid clashes with {stationaRy}
+
 ## Bug Fixes
 
 * `find_nearby_stations()`
@@ -12,7 +16,7 @@
   
   * Fixes the messages when stations aren't found nearby. In certain cases the message indicated that a `station_code` was used when `latitude` and `longitude` had been provided and _vice versa_.
   
-  * Fixes bug when there are no stations that meet the criteria set by the user and an error was emitted that "x" must be a data.table object. The function now simply errors if there are no stations from either API that meet the criteria.
+  * Fixes bug when there are no stations that meet the criteria set by the user and an error was emitted that "x" must be a {data.table} object. The function now simply errors if there are no stations from either API that meet the criteria.
 
 * `get_dpird_extremes()`
 
@@ -20,15 +24,14 @@
 
   * Fixes example for `get_dpird_extremes()` that didn't follow the package's standard use of "your_api_key" for the `api_key` value in the example.
 
-* Other
 
-  * Fixes the re-export of `terra::plot()` where checks indicated that it does not have a documented return value or examples.
+* Fixes the re-export of `terra::plot()` where checks indicated that it does not have a documented return value or examples.
 
-  * Fixes an issue where the package data was not available when this package was called by another, _e.g._, {extractOz} used `get_patched_point()` but could not validate the user-input `values` because `silo_daily_values` could not be found.
+* Fixes an issue where the package data was not available when this package was called by another, _e.g._, {extractOz} used `get_patched_point()` but could not validate the user-input `values` because `silo_daily_values` could not be found.
 
-  * Fixes bug where results from `get_dpird_summaries()` were truncated to only a few of the last results requested rather than the full set from start date to end date.
+* Fixes bug where results from `get_dpird_summaries()` were truncated to only a few of the last results requested rather than the full set from start date to end date.
 
-  * Fixes [bug](https://github.com/DPIRD-FSI/weatherOz/issues/38) where the DPIRD API would not properly respond to requests from Windows OS machines.
+* Fixes [bug](https://github.com/DPIRD-FSI/weatherOz/issues/38) where the DPIRD API would not properly respond to requests from Windows OS machines.
 
 # weatherOz 0.0.1
 
