+++ man/get_stations_metadata.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_station_metadata.R
+% Please edit documentation in R/get_stations_metadata.R
 \name{get_stations_metadata}
 \alias{get_stations_metadata}
 \title{Get Weather Station Metadata for Both DPIRD and SILO Weather Stations}
