### 2024-02-02 b92c6e3
Fix bug checking inputs for interval
