### 2024-02-03 1e49b9e
Fixes default value for DPIRD summaries interval and tests
