### 2024-02-03 03050b8
Reknit vignettes to ensure proper functionality reflected in documentation
