+++ vignettes/weatherOz_for_SILO.Rmd
@@ -105,71 +105,149 @@ library(weatherOz)
     api_key = Sys.getenv("SILO_API_KEY")
   )
 )
-#>      longitude latitude  year month   day       date air_tmax air_tmax_source air_tmin air_tmin_source  elev_m et_morton_actual
-#>          <num>    <num> <num> <num> <int>     <Date>    <num>           <int>    <num>           <int>  <char>            <num>
-#>   1:       150      -27  2022     1     1 2022-01-01     29.3              25     18.8              25 274.9 m              6.5
-#>   2:       150      -27  2022     1     2 2022-01-02     32.4              25     16.8              25 274.9 m              6.9
-#>   3:       150      -27  2022     1     3 2022-01-03     34.6              25     18.2              25 274.9 m              5.3
-#>   4:       150      -27  2022     1     4 2022-01-04     35.0              25     18.0              25 274.9 m              6.1
-#>   5:       150      -27  2022     1     5 2022-01-05     35.5              25     20.4              25 274.9 m              6.4
-#>  ---                                                                                                                           
-#> 758:       150      -27  2024     1    28 2024-01-28     33.1              25     23.4              25 274.9 m              4.2
-#> 759:       150      -27  2024     1    29 2024-01-29     32.4              25     25.2              25 274.9 m              4.4
-#> 760:       150      -27  2024     1    30 2024-01-30     30.6              25     23.4              25 274.9 m              3.5
-#> 761:       150      -27  2024     1    31 2024-01-31     33.1              25     21.6              25 274.9 m              6.4
-#> 762:       150      -27  2024     2     1 2024-02-01     33.3              75     21.3              25 274.9 m              4.2
-#>      et_morton_actual_source et_morton_potential et_morton_potential_source et_morton_wet et_morton_wet_source et_short_crop
-#>                        <int>               <num>                      <int>         <num>                <int>         <num>
-#>   1:                      26                 7.2                         26           6.8                   26           5.4
-#>   2:                      26                 8.9                         26           7.9                   26           6.6
-#>   3:                      26                10.7                         26           8.0                   26           7.4
-#>   4:                      26                10.2                         26           8.1                   26           7.3
-#>   5:                      26                10.4                         26           8.4                   26           7.4
-#>  ---                                                                                                                        
-#> 758:                      26                 6.3                         26           5.2                   26           4.4
-#> 759:                      26                 5.9                         26           5.1                   26           4.2
-#> 760:                      26                 4.9                         26           4.2                   26           3.5
-#> 761:                      26                 9.3                         26           7.9                   26           6.6
-#> 762:                      26                 8.9                         26           6.6                   26           6.1
-#>      et_short_crop_source et_tall_crop et_tall_crop_source evap_comb evap_comb_source evap_morton_lake evap_morton_lake_source
-#>                     <int>        <num>               <int>     <num>            <int>            <num>                   <int>
-#>   1:                   26          6.2                  26       4.9               25              7.0                      26
-#>   2:                   26          7.9                  26       9.6               25              8.3                      26
-#>   3:                   26          9.2                  26       8.2               25              8.4                      26
-#>   4:                   26          9.0                  26       8.6               25              8.5                      26
-#>   5:                   26          9.0                  26       6.3               25              8.8                      26
-#>  ---                                                                                                                          
[…]
