+++ vignettes/weatherOz_for_DPIRD.Rmd
@@ -158,32 +158,71 @@ library(weatherOz)
   )
 )
 #> Key: <station_code>
-#>       station_code           date_time air_temperature relative_humidity soil_temperature solar_irradiance rainfall dew_point
-#>             <fctr>              <POSc>           <num>             <num>            <num>            <int>    <int>     <num>
-#>    1:           NO 2024-02-01 12:47:00            44.8               8.8             49.5             1100        0       4.4
-#>    2:           NO 2024-02-01 12:48:00            44.9               8.6             49.6             1101        0       4.1
-#>    3:           NO 2024-02-01 12:49:00            45.1               8.8             49.7             1099        0       4.6
-#>    4:           NO 2024-02-01 12:50:00            45.3               8.5             49.7             1104        0       4.2
-#>    5:           NO 2024-02-01 12:51:00            45.0               8.6             49.8             1100        0       4.2
-#>   ---                                                                                                                        
-#> 1430:           NO 2024-02-02 12:36:00            42.7               9.4             49.4             1105        0       3.8
-#> 1431:           NO 2024-02-02 12:37:00            42.9               9.4             49.5             1103        0       3.9
-#> 1432:           NO 2024-02-02 12:38:00            43.2               9.3             49.6             1101        0       4.0
-#> 1433:           NO 2024-02-02 12:39:00            43.6               8.9             49.6             1104        0       3.6
-#> 1434:           NO 2024-02-02 12:40:00            43.6               8.7             49.7             1102        0       3.3
-#>       wet_bulb wind_height wind_avg_speed wind_avg_direction_compass_point wind_avg_direction_degrees wind_min_speed wind_max_speed
-#>          <num>       <int>          <num>                           <char>                      <int>          <num>          <num>
-#>    1:     19.1           3           4.68                               NE                         37          0.648           8.21
-#>    2:     19.0           3           1.66                               NE                         40          0.144           6.19
-#>    3:     19.2           3           2.66                               NE                         43          0.144           5.69
-#>    4:     19.1           3           4.18                              ENE                         59          1.656           9.22
-#>    5:     19.1           3           1.66                              ENE                         75          0.648           3.17
-#>   ---                                                                                                                              
-#> 1430:     18.6           3           4.18                              WSW                        242          2.664           7.70
-#> 1431:     18.7           3           3.17                               SE                        146          1.656           4.68
-#> 1432:     18.8           3           3.17                                N                          2          2.160           5.18
-#> 1433:     18.7           3           3.67                               NE                         45          0.144           7.70
-#> 1434:     18.6           3           8.21                              NNW                        345          3.672          11.23
[…]
