+++ vignettes/weatherOz.Rmd
@@ -61,10 +61,14 @@ Northam <-
   )
 
 Northam
-#>    station_code station_name longitude  latitude  state elev_m                                                                owner
-#>          <fctr>       <char>     <num>     <num> <char>  <int>                                                               <char>
-#> 1:           NO      Northam  116.6942 -31.65161     WA    163 WA Department of Primary Industries and Regional Development (DPIRD)
-#> 2:           MK       Muresk  116.6913 -31.72772     WA    251 WA Department of Primary Industries and Regional Development (DPIRD)
+#>    station_code station_name longitude  latitude  state elev_m
+#>          <fctr>       <char>     <num>     <num> <char>  <int>
+#> 1:           NO      Northam  116.6942 -31.65161     WA    163
+#> 2:           MK       Muresk  116.6913 -31.72772     WA    251
+#>                                                                   owner
+#>                                                                  <char>
+#> 1: WA Department of Primary Industries and Regional Development (DPIRD)
+#> 2: WA Department of Primary Industries and Regional Development (DPIRD)
 #>    distance_km
 #>          <num>
 #> 1:        3.23
@@ -86,32 +90,58 @@ dpird_daily <- get_dpird_summaries(
 
 dpird_daily
 #> Key: <station_code>
-#>      station_code station_name longitude  latitude  year month   day       date air_tavg air_tmax       air_tmax_time air_tmin
-#>            <fctr>       <char>     <num>     <num> <int> <int> <int>     <Date>    <num>    <num>              <POSc>    <num>
-#>   1:           NO      Northam  116.6942 -31.65161  2022     1     1 2022-01-01     23.8     32.5 2022-01-01 17:02:00     16.7
-#>   2:           NO      Northam  116.6942 -31.65161  2022     1     2 2022-01-02     25.4     36.1 2022-01-02 16:58:00     14.4
-#>   3:           NO      Northam  116.6942 -31.65161  2022     1     3 2022-01-03     28.0     39.9 2022-01-03 14:42:00     16.6
-#>   4:           NO      Northam  116.6942 -31.65161  2022     1     4 2022-01-04     26.1     36.3 2022-01-04 15:15:00     17.3
-#>   5:           NO      Northam  116.6942 -31.65161  2022     1     5 2022-01-05     27.3     39.1 2022-01-05 16:19:00     16.1
-#>  ---                                                                                                                          
-#> 361:           NO      Northam  116.6942 -31.65161  2022    12    27 2022-12-27     23.0     40.2 2022-12-27 14:48:20     13.7
-#> 362:           NO      Northam  116.6942 -31.65161  2022    12    28 2022-12-28     22.2     40.5 2022-12-28 19:58:20     11.9
-#> 363:           NO      Northam  116.6942 -31.65161  2022    12    29 2022-12-29     26.0     40.2 2022-12-30 03:03:20     13.8
-#> 364:           NO      Northam  116.6942 -31.65161  2022    12    30 2022-12-30     26.6     39.2 2022-12-30 13:50:50     14.5
-#> 365:           NO      Northam  116.6942 -31.65161  2022    12    31 2022-12-31     22.7     31.1 2022-12-31 15:45:50     16.2
-#>            air_tmin_time rainfall rh_avg rh_tmax        rh_tmax_time rh_tmin        rh_tmin_time
-#>                   <POSc>    <num>  <num>   <num>              <POSc>   <num>              <POSc>
-#>   1: 2022-01-01 05:17:00        0   39.9    71.5 2022-01-01 04:10:00    14.5 2022-01-01 18:03:00
[…]

[…]
