### 2024-02-09 #598 @ropensci-review-bot
Approved! Thanks @bozaah for submitting and @JanLauGe, @rogerssam for your reviews! :grin:
