### 2024-03-08 #598 @bozaah
Hi @ldecicco-USGS, I have made the transfer and tagged the bot to complete the process. We will work through the other items in the checklist next. We have also drafted a JOSS manuscript for submission, where can we find the next steps in the process please? Lastly, yhank you again for supporting us during the review and our appreciation to @JanLauGe and @rogerssam for their time and input improving our package.
