### 2024-03-08 #598 @ldecicco-USGS
All I know is:
https://joss.readthedocs.io/en/latest/submitting.html

And mention in the submission you've gone through an rOpenSci review, I'd put a link to this github issue. I'll double check there aren't more specific rOpenSci recommendations.
